% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-attach.R
\name{attach_mimic}
\alias{attach_mimic}
\title{Data attach utilities}
\usage{
attach_mimic(demo = FALSE, dir = if (demo) data_dir("mimic-demo") else
  data_dir("mimic-data"), cfg = if (demo) get_config("mimic-demo") else
  get_config("mimic-setup"), env = if (demo) mimic_demo else mimic)
}
\description{
Data attach utilities
}
