% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-env.R, R/setup-attach.R
\docType{data}
\name{data}
\alias{data}
\alias{aux}
\alias{attach_source}
\alias{attach_source.character}
\alias{attach_source.list}
\alias{attach_source.src_config}
\alias{mimic}
\alias{mimic_demo}
\alias{eicu}
\alias{eicu_demo}
\alias{hirid}
\title{Data attach utilities}
\format{
An object of class \code{environment} of length 5.

An object of class \code{environment} of length 5.
}
\usage{
data

aux

attach_source(x, ...)

\method{attach_source}{character}(x, file = NULL, ...)

\method{attach_source}{list}(x, ...)

\method{attach_source}{src_config}(x, dir = NULL, assign_env = .GlobalEnv, ...)
}
\arguments{
\item{x}{Data source to attach}

\item{...}{Forwarded to further calls to \code{attach_source()}}

\item{file}{File name string pointing to a non-default source configuration
file}

\item{dir}{Directory used to look for \code{\link[fst:fst]{fst::fst()}} files; \code{NULL} calls
\code{\link[=data_dir]{data_dir()}} using the source name as \code{subdir} argument}

\item{assign_env}{Environment in which the data source will become available}
}
\description{
Attaching a data source sets up an environment containing \code{\link[prt:new_prt]{prt::new_prt()}}
files using \code{\link[base:delayedAssign]{base::delayedAssign()}}.
}
\keyword{datasets}
