% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-attach.R
\docType{data}
\name{attach_src}
\alias{attach_src}
\alias{attach_src.src_cfg}
\alias{attach_src.character}
\alias{setup_src_env}
\alias{setup_src_env.src_cfg}
\alias{data}
\alias{mimic}
\alias{mimic_demo}
\alias{eicu}
\alias{eicu_demo}
\alias{hirid}
\alias{get_data_src}
\alias{get_src_env}
\alias{get_src_env.character}
\alias{get_src_env.default}
\alias{get_src_env.data_src}
\title{Data attach utilities}
\format{
An object of class \code{environment} of length 5.
}
\usage{
attach_src(x, ...)

\method{attach_src}{src_cfg}(x, dir = src_data_dir(x), assign_env = .GlobalEnv, ...)

\method{attach_src}{character}(x, name = "data-sources", file = NULL, ...)

setup_src_env(x, env, dir = NULL)

\method{setup_src_env}{src_cfg}(x, env, dir = src_data_dir(x))

data

get_data_src(tbl, src)

get_src_env(x)

\method{get_src_env}{character}(x)

\method{get_src_env}{default}(x)

\method{get_src_env}{data_src}(x)
}
\arguments{
\item{x}{Data source to attach}

\item{...}{Forwarded to further calls to \code{attach_src()}}

\item{dir}{Directory used to look for \code{\link[fst:fst]{fst::fst()}} files; \code{NULL} calls
\code{\link[=data_dir]{data_dir()}} using the source name as \code{subdir} argument}

\item{assign_env}{Environment in which the data source will become available}

\item{name}{String valued name of a config file which will be looked up in
the default config directors}

\item{file}{Full file name to load}

\item{env}{Environment where data proxy objects are created}

\item{tbl}{String valued table name}

\item{src}{Object passed to \code{get_src_env()} or an environment}
}
\description{
Attaching a data source sets up an environment containing \code{\link[prt:new_prt]{prt::new_prt()}}
files using \code{\link[base:delayedAssign]{base::delayedAssign()}}.
}
\keyword{datasets}
