% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-callback.R
\name{transform_fun}
\alias{transform_fun}
\alias{binary_op}
\alias{comp_na}
\alias{apply_map}
\alias{convert_unit}
\title{Callback utilities}
\usage{
transform_fun(fun, ...)

binary_op(op, y)

comp_na(op, y)

apply_map(map)

convert_unit(rgx, fun, new, ignore_case = TRUE, ...)
}
\arguments{
\item{fun}{Function(s) used for transforming matching values}

\item{...}{Further arguments passed to downstream function}

\item{op}{Function taking two arguments, such as \code{+}}

\item{y}{Value passed as second argument to function \code{op}}

\item{map}{Named atomic vector used for mapping a set of values (the names
of \code{map}) to a different set (the values of \code{map})}

\item{rgx}{Regular expression(s) used for identifying observations based on their current unit of measurement}

\item{new}{Name(s) of transformed units}

\item{ignore_case}{Forwarded to \code{\link[base:grep]{base::grep()}}}
}
\description{
Utilities for creating callback functions to be used for concept loading.
}
