% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{new_col_cfg}
\alias{new_col_cfg}
\alias{is_col_cfg}
\alias{as_col_cfg}
\alias{as_col_cfg.col_cfg}
\alias{as_col_cfg.list}
\alias{as_col_cfg.src_cfg}
\alias{as_col_cfg.src_tbl}
\title{Default columns}
\usage{
new_col_cfg(table, id, index, val, unit, time, src)

is_col_cfg(x)

as_col_cfg(x, ...)

\method{as_col_cfg}{col_cfg}(x, ...)

\method{as_col_cfg}{list}(x, src, ...)

\method{as_col_cfg}{src_cfg}(x, ...)

\method{as_col_cfg}{src_tbl}(x, ...)
}
\arguments{
\item{table}{Table name (string)}

\item{x}{Object to test/coerce}

\item{defaults}{Named list containing column names}

\item{times}{Character vector enumerating columns that contain time
information}
}
\description{
A table can have some of its columns marked as default columns for the
following concepts:
}
\details{
\itemize{
\item \code{id_col}: column will be used for as id for \code{icu_tbl} objects
\item \code{index_col}: column represents a timestamp variable and will be use as
such for \code{ts_tbl} objects
\item \code{val_col}: column contains the measured variable of interest
\item \code{unit_col}: column specifies the unit of measurement in the corresponding
\code{val_col}
}
}
\keyword{internal}
