% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{new_col_defaults}
\alias{new_col_defaults}
\alias{is_col_defaults}
\alias{as_col_defaults}
\alias{as_col_defaults.col_defaults}
\alias{as_col_defaults.list}
\alias{as_col_defaults.src_config}
\title{Default columns}
\usage{
new_col_defaults(table, cfg)

is_col_defaults(x)

as_col_defaults(x)

\method{as_col_defaults}{col_defaults}(x)

\method{as_col_defaults}{list}(x)

\method{as_col_defaults}{src_config}(x)
}
\arguments{
\item{table}{Table name (string)}

\item{cfg}{Named list containing column names}

\item{x}{Object to test/coerce}
}
\description{
A table can have some of its columns marked as default columns for the
following concepts:
}
\details{
\itemize{
\item \code{id_col}: column will be used for as id for \code{icu_tbl} objects
\item \code{index_col}: column represents a timestamp variable and will be use as
such for \code{ts_tbl} objects
\item \code{val_col}: column contains the measured variable of interest
\item \code{unit_col}: column specifies the unit of measurement in the corresponding
\code{val_col}
}
}
\keyword{internal}
