% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-utils.R
\name{col_renamer}
\alias{col_renamer}
\title{Internal utilities for ICU data classes}
\usage{
col_renamer(x, new, old = colnames(x), skip_absent = FALSE, by_ref = FALSE)
}
\arguments{
\item{x}{Object to query}

\item{new, old}{Replacement names and existing column names for renaming
columns}

\item{skip_absent}{Logical flag for ignoring non-existent column names}

\item{by_ref}{Logical flag indicating whether to perform the operation by
reference}
}
\description{
Internal utilities for ICU data classes
}
\keyword{internal}
