% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R
\name{load_dictionary}
\alias{load_dictionary}
\title{Load concept dictionaries}
\usage{
load_dictionary(
  src = NULL,
  concepts = NULL,
  name = "concept-dict",
  file = NULL
)
}
\arguments{
\item{src}{\code{NULL} or the name of a data source}

\item{concepts}{A character vector used to subset the concept dictionary or
\code{NULL} indicating no subsetting}

\item{name}{Name of the dictionary to be read}

\item{file}{File name of the dictionary}
}
\description{
Data can be specified in JSON format and parsed into \code{concept} objects.
}
