% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-attach.R
\docType{data}
\name{attach_mimic}
\alias{attach_mimic}
\alias{attach_eicu}
\alias{data}
\alias{mimic}
\alias{mimic_demo}
\alias{eicu}
\alias{eicu_demo}
\alias{attach_datasource}
\alias{setup_proxy_env}
\title{Data attach utilities}
\format{An object of class \code{environment} of length 4.}
\usage{
attach_mimic(
  demo = FALSE,
  dir = mimic_data_dir(demo),
  config = mimic_config(demo),
  envir = if (demo) mimic_demo else mimic
)

attach_eicu(
  demo = FALSE,
  dir = eicu_data_dir(demo),
  config = eicu_config(demo),
  envir = if (demo) eicu_demo else eicu
)

data

mimic

mimic_demo

eicu

eicu_demo

attach_datasource(dir, config)

setup_proxy_env(config, envir = new.env())
}
\arguments{
\item{demo}{Logical flag indicating whether to import the respective demo
of full dataset. Only used for determining the correct default arguments
passed on to \code{attach_datasource()}.}

\item{dir}{Data directory where the imported \code{.fst} files are located (see
\code{\link[=import_datasource]{import_datasource()}} for further information).}

\item{config}{Import configuration specified as nested list (see
\code{\link[=import_datasource]{import_datasource()}} for further information).}

\item{envir}{(Optional) environment that serves as proxy for the environment
holding references to tables. Passing \code{NULL} foregoes setting up this
forwarding (implemented as active bindings, see
\code{\link[base:makeActiveBinding]{base::makeActiveBinding()}}).}
}
\description{
Data attach utilities
}
\keyword{datasets}
