% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R, R/generics.R
\name{new_dictionary}
\alias{new_dictionary}
\alias{dictionary}
\alias{is_dictionary}
\alias{read_dictionary}
\alias{as_dictionary}
\title{Concept dictionary}
\usage{
new_dictionary(concepts)

dictionary(...)

is_dictionary(x)

read_dictionary(name = "concept-dict", file = NULL, ...)

as_dictionary(x)
}
\arguments{
\item{concepts}{A vector of concept objects}

\item{...}{Wrapped by \code{\link[base:list]{base::list()}} and forwarded to \code{\link[=as_concept]{as_concept()}}}

\item{x}{A potential \code{dictionary} object}

\item{name}{Name of the dictionary to be read}

\item{file}{File name of the dictionary}
}
\description{
Multiple concepts, each specifying data items in several sources, together
form a dictionary.
}
