% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{import_mimic}
\alias{import_mimic}
\alias{import_eicu}
\title{Data import utilities}
\usage{
import_mimic(demo = FALSE, dir = if (demo) data_dir("mimic-demo") else
  data_dir("mimic-data"), config = if (demo) get_config("mimic-demo")
  else get_config("mimic-setup"), cleanup = TRUE)

import_eicu(demo = FALSE, dir = if (demo) data_dir("eicu-demo") else
  data_dir("eicu-data"), config = if (demo) get_config("eicu-demo") else
  get_config("eicu-setup"), cleanup = TRUE)
}
\arguments{
\item{demo}{Logical flag for switching between default arguments for demo
or full data sources.}

\item{dir}{The directory where the data was downloaded to (see
\code{\link[=download_mimic]{download_mimic()}}).}

\item{config}{Import configuration specified as nested list.}

\item{cleanup}{Logical flag, if \code{TRUE} the original CSV data is deleted
after successful import.}
}
\description{
In order to speed up data access operations, \code{sepsr} does not directly use
the PhysioNet provided CSV files, but converts all data to \code{\link[fst:fst]{fst::fst()}}
format, which allows for random row and column access. Large tables are
split into chunks in order to keep memory requirements under control.
}
\details{
A nested list supplied as \code{config} argument controls how the CSV data is
parsed and ingested. A the top level, elements \code{name} (string), \code{base_url}
(string), \code{version} (string) and \code{tables} (list) are expected. For every
file to be read in, an entry in \code{tables} with the name of the corresponding
file is assumed, requiring a list named \code{col_spec} and optionally a list
named \code{partitioning}.

Every \code{col_spec} node again is expected to be a named list with names
corresponding to column names in the CSV and list valued entries. Finally,
a list node defining a single column is a named list with a required,
string-valued \code{type} entry which is interpreted as \code{readr} column
specification function (see \code{\link[readr:cols]{readr::cols()}}). Further list entries will be
passed as arguments to the column specification function. A column defined
as\preformatted{list(type = "col_datetime", format = "\%Y-\%m-\%d \%H:\%M:\%S")
}

will therefore result in a \code{readr} column specification like\preformatted{col_datetime(format = "\%Y-\%m-\%d \%H:\%M:\%S")
}

Entries controlling the partitioning of tables, specified as \code{partitioning}
list nodes are expected to be named lists of length 1, where the name
corresponds to the column used for partitioning and partitions are defined
as integer vector which is passed to \code{\link[base:findInterval]{base::findInterval()}}. This implies
that partitions can only be created on integer valued columns.

Putting all this together, the expected structure of \code{config} is as\preformatted{list(
  name = "some_name",
  base_url = "some_url",
  version = "major.minor",
  tables = list(
    file_name_table_1 = list(
      col_spec = list(
        colname_1 = list(type = "col_*"),
        colname_2 = list(type = "col_*"),
        ...
      )
    ),
    file_name_table_2 = list(
      col_spec = list(...),
      partitioning = list(colname_i = c(0L, ))
    ),
    ...
  )
)
}

Configurations for the MIMIC-III and eICU databases are available in
\code{extdata/config} as JSON files and can be read in using \code{\link[=get_config]{get_config()}}.
}
\examples{
\dontrun{

dir <- tempdir()
list.files(dir)

download_mimic(demo = TRUE, dest = dir)
list.files(dir)

import_mimic(demo = TRUE, dir = dir)
list.files(dir)

unlink(dir, recursive = TRUE)

}

}
