% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-src.R
\docType{data}
\name{data}
\alias{data}
\alias{mimic}
\alias{mimic_demo}
\alias{eicu}
\alias{eicu_demo}
\alias{hirid}
\title{Data}
\format{
An object of class \code{environment} of length 5.
}
\usage{
data
}
\description{
The \href{https://lcp.mit.edu/index.html}{Laboratory for Computational Physiology } (LCP) at MIT hosts several large-scale
databases of hospital intensive care units (ICUs), two of which can be
either downloaded in full or as demo subsets using utilities provided by
\code{ricu}, while a third data set is available only in full. While demo data
sets are publicly available, full download requires credentialed access
which can be gained by applying for up an account with \href{https://physionet.org}{PhysioNet }.
}
\section{MIMIC-III}{

The Medical Information Mart for Intensive Care
(\href{https://physionet.org/content/mimiciii/}{MIMIC}) database holds
detailed clinical data from roughly 60,000 patient stays in Beth Israel
Deaconess Medical Center (BIDMC) intensive care units between 2001 and 2012.
The database includes information such as demographics, vital sign
measurements made at the bedside (~1 data point per hour), laboratory test
results, procedures, medications, caregiver notes, imaging reports, and
mortality (both in and out of hospital). For further information, please
refer to the \href{https://mimic.physionet.org/about/mimic}{MIMIC-III documentation }.

The corresponding
\href{https://physionet.org/content/mimiciii-demo}{demo dataset}
contains the full data of a randomly selected subset of 100 patients from
the patient cohort with conformed in-hospital mortality. The only notable
data omission is the \code{noteevents} table, which unstructured text reports on
patients.

More recently, Philips Healthcare and LCP began assembling the eICU
Collaborative Research Database as a multi-center resource for ICU data.
Combining data of several critical care units throughout the continental
United States from the years 2014 and 2015, this database contains
deidentified health data associated with over 200,000 admissions, including
vital sign measurements, care plan documentation, severity of illness
measures, diagnosis information, and treatment information. For further
information, please refer to the \href{https://eicu-crd.mit.edu/about/eicu}{eICU documentation }.
}

\references{
MIMIC-III, a freely accessible critical care database. Johnson AEW, Pollard
TJ, Shen L, Lehman L, Feng M, Ghassemi M, Moody B, Szolovits P, Celi LA,
and Mark RG. Scientific Data (2016). DOI: 10.1038/sdata.2016.35.

The eICU Collaborative Research Database, a freely available multi-center
database for critical care research. Pollard TJ, Johnson AEW, Raffa JD,
Celi LA, Mark RG and Badawi O. Scientific Data (2018). DOI:
http://dx.doi.org/10.1038/sdata.2018.178.
}
\keyword{datasets}
