% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{id_origin}
\alias{id_origin}
\alias{id_origin.id_cfg}
\alias{id_origin.eicu_ids}
\alias{id_origin.default}
\alias{id_windows}
\alias{id_windows.mimic_ids}
\alias{id_windows.eicu_ids}
\alias{id_windows.hirid_ids}
\alias{id_windows.default}
\title{Data loading utilities}
\usage{
id_origin(x, ...)

\method{id_origin}{id_cfg}(x, id, env = as_src_env(x), ...)

\method{id_origin}{eicu_ids}(x, ...)

\method{id_origin}{default}(x, id, env = as_src_env(x), ...)

id_windows(x, ...)

\method{id_windows}{mimic_ids}(x, env = as_src_env(x), ...)

\method{id_windows}{eicu_ids}(x, env = as_src_env(x), ...)

\method{id_windows}{hirid_ids}(x, env = as_src_env(x), ...)

\method{id_windows}{default}(x, env = as_src_env(x), ...)
}
\arguments{
\item{x}{Object to dispatch on}

\item{...}{Generic consistency}

\item{id}{ID name for which to return origin times}

\item{env}{Data environment in which to look for tables to construct helper
tables}
}
\description{
Two important tools for smoothing out differences among used datasets are
\code{id_origin()} which returns origin times for a given ID and \code{id_windows()}
which defines stay windows and relationships between ID systems.
}
