% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{default_col}
\alias{default_col}
\title{Get default columns}
\usage{
default_col(x, type)
}
\arguments{
\item{x}{Object used for dispatch}

\item{type}{The column type, e.g. \code{id}, \code{index}, \code{unit}, \code{value}}
}
\description{
For a table, query the default columns as specified by an \code{id_cols}
(\code{\link[=new_id_cols]{new_id_cols()}}) or \code{col_defaults} (\code{\link[=new_col_defaults]{new_col_defaults()}}) object.
}
