% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-download.R
\name{download_src}
\alias{download_src}
\alias{download_src.src_cfg}
\alias{download_src.character}
\title{Data download utilities}
\usage{
download_src(x, ...)

\method{download_src}{src_cfg}(
  x,
  dir = src_data_dir(x),
  tables = NULL,
  force = FALSE,
  user = NULL,
  pass = NULL,
  ...
)

\method{download_src}{character}(x, name = "data-sources", file = NULL, ...)
}
\arguments{
\item{x}{Object specifying the source configuration}

\item{...}{Generic consistency}

\item{dir}{Destination directory where the downloaded data is written to.}

\item{tables}{Character vector specifying the tables to download. If
\code{NULL}, all available tables are downloaded.}

\item{force}{Logical flag; if \code{TRUE}, existing data will be re-downloaded}

\item{user, pass}{Physionet credentials; if \code{NULL} and environment
variables \code{RICU_PHYSIONET_USER}/\code{RICU_PHYSIONET_PASS} are not set, user
input is required}

\item{name}{String valued name of a config file which will be looked up in
the default config directors}

\item{file}{Full file name to load}
}
\description{
The \href{https://lcp.mit.edu/index.html}{Laboratory for Computational Physiology } (LCP) at MIT hosts two large-scale
databases of hospital intensive care units (ICUs), both of which can be
either downloaded in full or as demo subsets. While both databases are
publicly available, full download requires credentialed access which can be
gained by applying for up an account with \href{https://physionet.org}{PhysioNet }.
}
\details{
The Medical Information Mart for Intensive Care (MIMIC) database holds
detailed clinical data from roughly 60,000 patient stays in Beth Israel
Deaconess Medical Center (BIDMC) intensive care units between 2001 and 2012.
The database includes information such as demographics, vital sign
measurements made at the bedside (~1 data point per hour), laboratory test
results, procedures, medications, caregiver notes, imaging reports, and
mortality (both in and out of hospital). For further information, please
refer to the \href{https://mimic.physionet.org/about/mimic}{MIMIC-III documentation }.

More recently, Philips Healthcare and LCP began assembling the eICU
Collaborative Research Database as a multi-center resource for ICU data.
Combining data of several critical care units throughout the continental
United States from the years 2014 and 2015, this database contains
deidentified health data associated with over 200,000 admissions, including
vital sign measurements, care plan documentation, severity of illness
measures, diagnosis information, and treatment information. For further
information, please refer to the \href{https://eicu-crd.mit.edu/about/eicu}{eICU documentation }.

Physionet credentials can either be entered in an interactive session,
passed as function arguments \code{user}/\code{pass} or as environment
variables \code{RICU_PHYSIONET_USER}/\code{RICU_PHYSIONET_PASS}. If the openssl
package is available, SHA256 hashes of downloaded files are verified using
\code{\link[openssl:hash]{openssl::sha256()}}.
}
\examples{
\dontrun{

dir <- tempdir()
list.files(dir)

download_datasource("mimic_demo", dir = dir)
list.files(dir)

unlink(dir, recursive = TRUE)

}

}
\references{
MIMIC-III, a freely accessible critical care database. Johnson AEW, Pollard
TJ, Shen L, Lehman L, Feng M, Ghassemi M, Moody B, Szolovits P, Celi LA,
and Mark RG. Scientific Data (2016). DOI: 10.1038/sdata.2016.35.

The eICU Collaborative Research Database, a freely available multi-center
database for critical care research. Pollard TJ, Johnson AEW, Raffa JD,
Celi LA, Mark RG and Badawi O. Scientific Data (2018). DOI:
http://dx.doi.org/10.1038/sdata.2018.178.
}
