% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{default_id_col}
\alias{default_id_col}
\alias{default_index_col}
\alias{default_val_col}
\alias{default_unit_col}
\title{Get default columns}
\usage{
default_id_col(table, ...)

default_index_col(table, ...)

default_val_col(table, ...)

default_unit_col(table, ...)
}
\arguments{
\item{table}{String valued table name}

\item{...}{Passed to further methods}
}
\description{
For a table, query the default columns as specified by an \code{id_cols}
(\code{\link[=new_id_cols]{new_id_cols()}}) or \code{col_defaults} (\code{\link[=new_col_defaults]{new_col_defaults()}}) object.
}
