% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-icu.R, R/tbl-id.R, R/tbl-ts.R
\name{is_icu_tbl}
\alias{is_icu_tbl}
\alias{id_tbl}
\alias{as_id_tbl}
\alias{as_id_tbl.default}
\alias{is_id_tbl}
\alias{ts_tbl}
\alias{as_ts_tbl}
\alias{as_ts_tbl.default}
\alias{is_ts_tbl}
\alias{as_id_tbl.ts_tbl}
\title{Tabular ICU data}
\usage{
is_icu_tbl(x)

id_tbl(..., id, id_opts = NULL)

as_id_tbl(x, ...)

\method{as_id_tbl}{default}(x, id, id_opts = NULL, ...)

is_id_tbl(x)

ts_tbl(..., id, id_opts = NULL, index = NULL, interval = hours(1L))

as_ts_tbl(x, ...)

\method{as_ts_tbl}{default}(x, id, id_opts = NULL, index = NULL, interval = hours(1L), ...)

is_ts_tbl(x)

\method{as_id_tbl}{ts_tbl}(x, ...)
}
\arguments{
\item{x}{Object to test/coerce}

\item{...}{forwarded to \code{\link[data.table:data.table]{data.table::data.table()}} or generic consistency}

\item{id}{String valued column name used as \code{id} column}

\item{id_opts}{\code{NULL} or a named character vector indicating options to
which ids may be converted see \code{\link[=change_id]{change_id()}}}

\item{index}{Name of index column}

\item{interval}{Time spacing between rows}
}
\description{
Two objects central to \code{ricu}, \code{id_tbl} and \code{ts_tbl} (both inheriting
from \code{icu_tbl}) are used to represent tabular data.
}
