% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-class.R, R/config-utils.R
\name{new_id_cfg}
\alias{new_id_cfg}
\alias{is_id_cfg}
\alias{as_id_cfg}
\alias{as_id_cfg.id_cfg}
\alias{as_id_cfg.src_cfg}
\alias{as_id_cfg.src_env}
\alias{as_id_cfg.default}
\alias{id_var_opts}
\title{ID columns}
\usage{
new_id_cfg(
  name,
  src,
  id,
  pos,
  start = NA_character_,
  end = NA_character_,
  table = NA_character_,
  class_prefix = src
)

is_id_cfg(x)

as_id_cfg(x)

\method{as_id_cfg}{id_cfg}(x)

\method{as_id_cfg}{src_cfg}(x)

\method{as_id_cfg}{src_env}(x)

\method{as_id_cfg}{default}(x)

id_var_opts(x)
}
\arguments{
\item{x}{Object to test/coerce}

\item{cfg}{List containing string valued column names}
}
\description{
A set of id columns corresponding to \code{patient}, \code{hadm} (hospital stay) and
\code{icustay} (ICU stay) can be represented by an \code{id_cfg} object.
}
\keyword{internal}
