% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{new_id_cols}
\alias{new_id_cols}
\alias{is_id_cols}
\alias{get_id_cols}
\alias{get_id_cols.id_cols}
\alias{get_id_cols.src_config}
\alias{get_id_cols.data_src}
\alias{get_id_cols.default}
\title{ID columns}
\usage{
new_id_cols(cfg, src)

is_id_cols(x)

get_id_cols(x, ...)

\method{get_id_cols}{id_cols}(x, id_name = NULL, ...)

\method{get_id_cols}{src_config}(x, ...)

\method{get_id_cols}{data_src}(x, ...)

\method{get_id_cols}{default}(x, id_name = NULL, ...)
}
\arguments{
\item{cfg}{List containing string valued column names}

\item{x}{Object to test/coerce}
}
\description{
A set of id columns corresponding to \code{patient}, \code{hadm} (hospital stay) and
\code{icustay} (ICU stay) can be represented by an \code{id_cols} object.
}
\keyword{internal}
