% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R
\name{prepare_query}
\alias{prepare_query}
\alias{try_add_vars}
\alias{get_itm_var}
\alias{set_callback}
\alias{do_callback}
\alias{do_itm_load}
\alias{n_tick}
\alias{set_target}
\alias{get_target}
\title{Internal utilities for \code{item}/\code{concept} objects}
\usage{
prepare_query(x)

try_add_vars(x, ...)

get_itm_var(x, var = NULL)

set_callback(x, fun)

do_callback(x, ...)

do_itm_load(x, id_type = "icustay", interval = hours(1L))

n_tick(x)

set_target(x, target)

get_target(x)
}
\arguments{
\item{x}{Object defining the row-subsetting}

\item{...}{Variable name}

\item{fun}{Callback function (passed as string)}

\item{id_type}{String specifying the patient id type to return}

\item{interval}{The time interval used to discretize time stamps with,
specified as \code{\link[base:difftime]{base::difftime()}} object}
}
\description{
Internal utilities for \code{item}/\code{concept} objects
}
\keyword{internal}
