% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-sofa.R
\encoding{UTF-8}
\name{sofa_score}
\alias{sofa_score}
\alias{sofa_window}
\alias{sofa_compute}
\title{SOFA score label}
\usage{
sofa_score(
  pafi,
  vent,
  plt,
  bili,
  map,
  norepi,
  epi,
  dopa,
  dobu,
  gcs,
  crea,
  urine24,
  ...,
  interval = NULL
)

sofa_window(
  tbl,
  pafi_win_fun = min_or_na,
  plt_win_fun = min_or_na,
  bili_win_fun = max_or_na,
  map_win_fun = min_or_na,
  dopa_win_fun = max_or_na,
  norepi_win_fun = max_or_na,
  dobu_win_fun = max_or_na,
  epi_win_fun = max_or_na,
  gcs_win_fun = min_or_na,
  crea_win_fun = max_or_na,
  urine_win_fun = min_or_na,
  explicit_wins = FALSE,
  worst_win_length = hours(24L)
)

sofa_compute(
  tbl,
  na_val = 0L,
  na_val_resp = na_val,
  na_val_plt = na_val,
  na_val_liver = na_val,
  na_val_cardio = na_val,
  na_val_cns = na_val,
  na_val_renal = na_val,
  impute_fun = NULL,
  keep_components = FALSE,
  by_ref = FALSE
)
}
\arguments{
\item{pafi, vent, plt, bili, map, norepi, epi, dopa, dobu, gcs, crea, urine24}{Data
input used for sofa score evaluation (\code{ts_tbl} objects, as produced by
\code{\link[=load_concepts]{load_concepts()}})}

\item{...}{Passed to \code{sofa_window()} or \code{sofa_compute()}}

\item{interval}{Time series interval (only used for checking consistency
of input data, \code{NULL} will use the interval of the first data object)}

\item{tbl}{Table holding SOFA covariates}

\item{pafi_win_fun, plt_win_fun, bili_win_fun, map_win_fun, dopa_win_fun, norepi_win_fun, dobu_win_fun, epi_win_fun, gcs_win_fun, crea_win_fun, urine_win_fun}{functions used to calculate worst values over windows}

\item{explicit_wins}{The default \code{FALSE} iterates over all time steps,
\code{TRUE} uses only the last time step per patient and a vector of times will
iterate over these explicit time points}

\item{worst_win_length}{Time-frame to look back and apply the \verb{*_win_fun}s}

\item{na_val}{Value to use for missing data}

\item{na_val_resp, na_val_plt, na_val_liver, na_val_cardio, na_val_cns, na_val_renal}{Feature-specific values to use in case of missing data; default is \code{na_val}}

\item{impute_fun}{Function used to impute missing data; default is NULL}

\item{keep_components}{Logical flag indicating whether to return individual
SOFA components as columns}

\item{by_ref}{Logical flag indicating whether to perform the operation by
reference}
}
\description{
The SOFA (Sequential Organ Failure Assessment) score is a commonly used
assessment tool used to track a patient's well-being in an ICU.
}
\details{
The SOFA score (Vincent et. al.) is evaluated as follows:\tabular{lllll}{
   \strong{SOFA score} \tab 1 \tab 2 \tab 3 \tab 4 \cr
   \strong{Respiration} PaO₂/FiO₂ [mmHg] \tab < 400 \tab < 300 \tab < 200 (with mech. vent.) \tab < 100 (with mech. vent.) \cr
   \strong{Coagulation} Platelets [⨉10³/mm³] \tab < 150 \tab < 100 \tab < 50 \tab < 20 \cr
   \strong{Liver} Bilirubin [mg/dl] ([μmol/μl]) \tab 1.2-1.9 (20-32) \tab 2.0-5.9 (33-101) \tab 6.0-11.9 (102-204) \tab > 12.0 (> 204) \cr
   \strong{Cardiovascular} Hypotension \tab MAP < 70 mmHg \tab Dopa ≤ 5 or dobu (any dose)ª \tab Dopa > 5 or epi ≤ 0.1 or norepi ≤ 0.1 \tab Dopa > 15 or epi > 0.1 or norepi > 0.1 \cr
   \strong{Central nervous system} Glasgow Coma Score \tab 13-14 \tab 10-12 \tab 6-9 \tab < 6 \cr
   \strong{Renal} Creatinine [mg/dl] ([μmol/μl]) or urine output [ml/day] \tab 1.2-1.9 (110-170) \tab 2.0-3.4 (171-299) \tab 3.5-4.9 (300-440) < 500 \tab > 5.0 (< 440) < 200 \cr
}


ªAdrenergic agents administered for at least 1h (doses given are in [μg/kg
⋅ min]; dopa: dopamine, dobu: dobutamine, epi: epinephrine, norepi:
norepinephrine)

For each component, the worst value over a window of length
\code{worst_win_length} (default \code{hours(24L)}) is taken.

The respiratory component of the SOFA score requires the evaluation of the
PaO₂/FiO₂ ratio. Often, however, both measures are not available at the
same time point. Therefore, PaO₂ and FiO₂ measurements are matched within a
window of size \code{pafi_win_length} (default \code{hours(2L)}).

The renal component of the SOFA score assigns scores of 3 and 4 in case of
a very low urine output. Since the information on urine output is not very
reliable at the start of the ICU stay, we do not evaluate urine outputs
before a time window of \code{urine_min_win} (default \code{hours(12L)}).

Mechanical ventilation is also part of the SOFA score. In some datasets,
ventilation start events are given, which cannot be matched to ventilation
end events specifically. In these cases, the duration of the ventilation
window is taken to be \code{vent_win_length} (default \code{hours(6L)}).
}
\references{
Vincent, J.-L., Moreno, R., Takala, J. et al. The SOFA (Sepsis-related Organ
Failure Assessment) score to describe organ dysfunction/failure. Intensive
Care Med 22, 707–710 (1996). https://doi.org/10.1007/BF01709751
}
