% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-load.R
\name{load_concepts}
\alias{load_concepts}
\alias{load_concept}
\title{Load concept data}
\usage{
load_concepts(concepts, aggregate = NA_character_, merge_data = TRUE, ...)

load_concept(concept, aggregate = NA_character_, na_rm = TRUE, ...)
}
\arguments{
\item{concepts}{Vector of \code{concept} objects to be loaded}

\item{aggregate}{Controls how data within concepts is aggregated}

\item{merge_data}{Logical flag, specifying whether to merge concepts into
wide format or return a list, each entry corresponding to a concept}

\item{...}{Passed to \code{load_concept()} and on to \code{\link[=load_items]{load_items()}}}

\item{concept}{Single \code{concept} object to be loaded}

\item{na_rm}{Logical flag, indicating whether to remove \code{NA} rows}
}
\description{
Data specified as \code{concept} objects is loaded, aggregated per combination
of id and time-step and merged into wide format.
}
