% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{data_ts}
\alias{data_ts}
\alias{data_ts_quo}
\alias{data_id}
\alias{data_id_quo}
\alias{data_tbl}
\alias{data_tbl_quo}
\title{Low level functions for loading data}
\usage{
data_ts(source, table, row_expr, ...)

data_ts_quo(
  source,
  table,
  row_quo = NULL,
  cols = NULL,
  id_col = default_id_col(config = cfg),
  time_col = default_time_col(table, config = cfg),
  interval = hours(1L),
  cfg = get_col_config(source, "all"),
  ...
)

data_id(source, table, row_expr, ...)

data_id_quo(
  source,
  table,
  row_quo = NULL,
  cols = NULL,
  id_col = default_id_col(config = cfg),
  interval = hours(1L),
  cfg = get_col_config(source, "all"),
  ...
)

data_tbl(source, table, row_expr, ...)

data_tbl_quo(
  source,
  table,
  row_quo = NULL,
  cols = NULL,
  interval = hours(1L),
  data_fun = get_col_config(source, "data_fun")
)
}
\arguments{
\item{source}{String specifying the data source}

\item{table}{String specifying the table from which to load data}

\item{row_expr}{Expression to be quoted and used to define a row subsetting;
\code{NULL} corresponds to no row subsetting}

\item{...}{Forwarded to the corresponding function that takes a quoted
expression and further to \code{change_id()}}

\item{row_quo}{Quoted expression used for defining a row subsetting; \code{NULL}
corresponds to no row subsetting}

\item{cols}{Character vector, specifying the column to return (in addition
to potential columns specified as \code{id_col} and \code{time_col}); \code{NULL}
corresponds to all columns}

\item{id_col}{The column defining the id of \code{ts_tbl} and \code{id_tbl} objects}

\item{time_col}{The column defining the index of \code{ts_tbl} objects}

\item{interval}{The time interval used to discretize time stamps with,
specified as \code{\link[base:difftime]{base::difftime()}} object}

\item{cfg}{Configuration list used for determining default options for
\code{id_col} and \code{time_col} arguments, id options as well as the \code{data_fun}
argument}

\item{data_fun}{Function used for loading the data; the default uses
\code{\link[prt:subset_quo]{prt::subset_quo()}} and for non-default values, a function with arguments
\code{table}, \code{row_quo}, \code{cols}, \code{interval}, and \code{source} is expected}
}
\description{
Depending on the desired return type (\code{ts_tbl}, \code{id_tbl} or plain
\code{data.table}) and on whether the row-subsetting argument is a quoted
expression (suffix \verb{_quo}), this set of functions provides the basic
mechanisms for loading data.
}
