% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-load.R
\name{load_dictionary}
\alias{load_dictionary}
\title{Load data from a dictionary}
\usage{
load_dictionary(
  source = NULL,
  concepts = NULL,
  dictionary = read_dictionary(),
  ...
)
}
\arguments{
\item{source, concepts}{A character vector, used to subset the \code{dictionary};
\code{NULL} means no subsetting}

\item{dictionary}{The dictionary to be used}

\item{...}{Passed to \code{\link[=load_concepts]{load_concepts()}}}
}
\description{
Using a data \code{dictionary} (or a subset thereof), data is loaded and pre-
processed as specified by the selected \code{concepts}.
}
