% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-load.R
\name{load_items}
\alias{load_items}
\alias{load_item}
\title{Load data items}
\usage{
load_items(items, units = NULL, ...)

load_item(
  item,
  unit = NULL,
  id_type = "hadm",
  patient_ids = NULL,
  interval = hours(1L),
  cfg = get_src_config(item)
)
}
\arguments{
\item{items}{Vector of data \code{item} objects}

\item{units}{Character vector of units (one per data item)}

\item{...}{Passed on to \code{\link[=load_item]{load_item()}}}

\item{item}{Single \code{item} object}

\item{unit}{String describing the unit of measurement of the given item}

\item{id_type}{String specifying the patient id type to return}

\item{patient_ids}{Optional vector of patient ids to subset the fetched data
with}

\item{interval}{The time interval used to discretize time stamps with,
specified as \code{\link[base:difftime]{base::difftime()}} object}

\item{cfg}{Configuration list used for determining default options for
\code{id_col} and \code{time_col} arguments, id options as well as the \code{data_fun}
argument}
}
\description{
Load data specified as data \code{item} objects.
}
