% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{load_id}
\alias{load_id}
\alias{load_id.src_tbl}
\alias{load_id.character}
\alias{load_id.default}
\alias{load_ts}
\alias{load_ts.src_tbl}
\alias{load_ts.character}
\alias{load_ts.default}
\title{Load data as \code{id_tbl} or \code{ts_tbl} objects}
\usage{
load_id(x, ...)

\method{load_id}{src_tbl}(
  x,
  rows,
  cols = colnames(x),
  id_var = id_vars(x),
  interval = hours(1L),
  time_vars = ricu::time_vars(x),
  ...
)

\method{load_id}{character}(x, src, ...)

\method{load_id}{default}(x, ...)

load_ts(x, ...)

\method{load_ts}{src_tbl}(
  x,
  rows,
  cols = colnames(x),
  id_var = id_vars(x),
  index_var = ricu::index_var(x),
  interval = hours(1L),
  time_vars = ricu::time_vars(x),
  ...
)

\method{load_ts}{character}(x, src, ...)

\method{load_ts}{default}(x, ...)
}
\arguments{
\item{x}{Object for which to load data}

\item{...}{Generic consistency}

\item{rows}{Expression used for row subsetting (NSE)}

\item{cols}{Character vector of column names}

\item{id_var}{The column defining the id of \code{ts_tbl} and \code{id_tbl} objects}

\item{interval}{The time interval used to discretize time stamps with,
specified as \code{\link[base:difftime]{base::difftime()}} object}

\item{time_vars}{Character vector enumerating the columns to be treated as
timestamps and thus returned as \code{\link[base:difftime]{base::difftime()}} vectors}

\item{src}{Passed to \code{\link[=as_src_tbl]{as_src_tbl()}} in order to determine the data source}

\item{index_var}{The column defining the index of \code{ts_tbl} objects}
}
\description{
Load data as \code{id_tbl} or \code{ts_tbl} objects
}
