% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-id.R, R/meta-tbl.R, R/meta-ts.R,
%   R/meta-utils.R, R/tbl-icu.R, R/tbl-ts.R
\name{set_id.id_meta}
\alias{set_id.id_meta}
\alias{index.tbl_index}
\alias{interval.tbl_index}
\alias{time_unit.tbl_index}
\alias{id.tbl_id}
\alias{id_opts.tbl_id}
\alias{set_id.tbl_id}
\alias{set_id_opts.tbl_id}
\alias{set_index.tbl_index}
\alias{set_interval.tbl_index}
\alias{set_time_unit.tbl_index}
\alias{index.ts_meta}
\alias{interval.ts_meta}
\alias{time_unit.ts_meta}
\alias{meta_cols.ts_meta}
\alias{set_id.ts_meta}
\alias{set_index.ts_meta}
\alias{set_interval.ts_meta}
\alias{set_time_unit.ts_meta}
\alias{id.tbl_meta}
\alias{id_opts.tbl_meta}
\alias{index}
\alias{id}
\alias{id_opts}
\alias{interval}
\alias{time_unit}
\alias{meta_cols}
\alias{set_index}
\alias{set_interval}
\alias{set_id}
\alias{set_id_opts}
\alias{set_time_unit}
\alias{id.icu_tbl}
\alias{id_opts.icu_tbl}
\alias{set_id.icu_tbl}
\alias{set_id_opts.icu_tbl}
\alias{index.icu_tbl}
\alias{units.icu_tbl}
\alias{meta_cols.icu_tbl}
\alias{data_cols}
\alias{data_unit}
\alias{time_step}
\alias{time_col}
\title{Utilities for querying and modifying table meta data}
\usage{
\method{set_id}{id_meta}(x, value)

\method{index}{tbl_index}(x)

\method{interval}{tbl_index}(x)

\method{time_unit}{tbl_index}(x)

\method{id}{tbl_id}(x)

\method{id_opts}{tbl_id}(x)

\method{set_id}{tbl_id}(x, value)

\method{set_id_opts}{tbl_id}(x, value)

\method{set_index}{tbl_index}(x, value)

\method{set_interval}{tbl_index}(x, value)

\method{set_time_unit}{tbl_index}(x, value)

\method{index}{ts_meta}(x)

\method{interval}{ts_meta}(x)

\method{time_unit}{ts_meta}(x)

\method{meta_cols}{ts_meta}(x)

\method{set_id}{ts_meta}(x, value)

\method{set_index}{ts_meta}(x, value)

\method{set_interval}{ts_meta}(x, value)

\method{set_time_unit}{ts_meta}(x, value)

\method{id}{tbl_meta}(x)

\method{id_opts}{tbl_meta}(x)

index(x)

id(x)

id_opts(x)

interval(x)

time_unit(x)

meta_cols(x)

set_index(x, value)

set_interval(x, value)

set_id(x, value)

set_id_opts(x, value)

set_time_unit(x, value)

\method{id}{icu_tbl}(x)

\method{id_opts}{icu_tbl}(x)

\method{set_id}{icu_tbl}(x, value)

\method{set_id_opts}{icu_tbl}(x, value)

\method{index}{icu_tbl}(x)

\method{units}{icu_tbl}(x)

\method{meta_cols}{icu_tbl}(x)

data_cols(x)

data_unit(x)

time_step(x)

time_col(x)
}
\arguments{
\item{x}{Object to query/modify}

\item{value}{New value to set}
}
\description{
Utilities for querying and modifying table meta data
}
