% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-class.R
\name{read_src_cfg}
\alias{read_src_cfg}
\title{Load configuration for a data source}
\usage{
read_src_cfg(src = NULL, name = "data-sources", file = NULL)
}
\arguments{
\item{src}{(Optional) name(s) of data sources used for subsetting}

\item{name}{String valued name of a config file which will be looked up in
the default config directors}

\item{file}{Full file name to load}
}
\description{
For a data source to become available to ricu, a JSON base configuration
file is required which is parsed into a \code{src_cfg} object (see
\code{\link[=new_src_cfg]{new_src_cfg()}}).
}
