% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricu.R
\docType{package}
\name{ricu-package}
\alias{ricu}
\alias{ricu-package}
\title{ricu: Tools for working with publicly available intensive care unit (ICU) data sets}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/septic-tank/ricu}
}

}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas.bennett@stat.math.ethz.ch}

Authors:
\itemize{
  \item Drago Plecko \email{drago.plecko@stat.math.ethz.ch}
  \item Ida-Fong Ukor \email{dr.ukor@gmail.com}
}

}
\keyword{internal}
