% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-sepsis3.R
\name{sepsis_3}
\alias{sepsis_3}
\alias{delta_cummin}
\alias{delta_start}
\alias{delta_min}
\title{Sepsis 3 label}
\usage{
sepsis_3(
  sofa,
  si,
  si_window = c("first", "last", "any"),
  delta_fun = delta_cummin,
  sofa_thresh = 2L
)

delta_cummin(x)

delta_start(x)

delta_min(x, shifts = seq.int(0L, 23L))
}
\arguments{
\item{sofa}{\code{ts_tbl} with a column \code{sofa_score}}

\item{si}{\code{ts_tbl} with columns \code{si_lwr} and \code{si_upr} defining windows for
which a suspected infection is valid}

\item{si_window}{Switch that can be used to filter SI windows}

\item{delta_fun}{Function used to determine the SOFA increase during an SI
window}

\item{sofa_thresh}{Required SOFA increase to trigger Sepsis 3}

\item{x}{Vector of SOFA scores}

\item{shifts}{Vector of time shifts (multiples of the current interval) over
which \code{\link[base:pmin]{base::pmin()}} is evaluated}
}
\description{
The sepsis 3 label consists of a suspected infection combined with an acute
increase in SOFA score.
}
