% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{stay_windows}
\alias{stay_windows}
\title{Stays}
\usage{
stay_windows(
  source,
  id_type = "icustay",
  win_type = "icustay",
  in_time = "intime",
  out_time = "outtime",
  interval = hours(1L)
)
}
\arguments{
\item{source}{String specifying the data source}

\item{id_type}{Type of ID all returned times are relative to}

\item{win_type}{Type of ID for which the in/out times is returned}

\item{in_time, out_time}{column names of the returned in/out times}

\item{interval}{The time interval used to discretize time stamps with,
specified as \code{\link[base:difftime]{base::difftime()}} object}
}
\description{
For a given ID type, get all stays with corresponding start and end times.
}
