% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{new_table_spec}
\alias{new_table_spec}
\alias{is_table_spec}
\alias{as_table_spec}
\alias{as_table_spec.table_spec}
\alias{as_table_spec.list}
\alias{get_table_specs}
\alias{table_names}
\alias{table_names.src_config}
\alias{table_names.list}
\alias{table_names.table_spec}
\alias{file_names}
\alias{file_names.src_config}
\alias{file_names.list}
\alias{file_names.table_spec}
\alias{n_partitions}
\alias{n_partitions.src_config}
\alias{n_partitions.list}
\alias{n_partitions.table_spec}
\alias{fst_names}
\alias{get_col_spec}
\alias{check_n_rows}
\alias{get_col_names}
\alias{get_part_fun}
\title{Table specification}
\usage{
new_table_spec(table, files, cols, nrow = NULL, partitioning = NULL)

is_table_spec(x)

as_table_spec(x)

\method{as_table_spec}{table_spec}(x)

\method{as_table_spec}{list}(x)

get_table_specs(x)

table_names(x)

\method{table_names}{src_config}(x)

\method{table_names}{list}(x)

\method{table_names}{table_spec}(x)

file_names(x)

\method{file_names}{src_config}(x)

\method{file_names}{list}(x)

\method{file_names}{table_spec}(x)

n_partitions(x)

\method{n_partitions}{src_config}(x)

\method{n_partitions}{list}(x)

\method{n_partitions}{table_spec}(x)

fst_names(x)

get_col_spec(x)

check_n_rows(x, comparison)

get_col_names(x)

get_part_fun(x, orig_names = FALSE)
}
\arguments{
\item{table}{String valued table name}

\item{cols}{List containing a list per column each holding string valued
entries \code{name} (column name as used by \code{ricu}), \code{col} (column name as used
in the raw data) and \code{spec} (name of \code{\link[readr:cols]{readr::cols()}} column specification).
Further entries will be passed as argument to the respective \code{readr} column
specification}

\item{nrow}{A count indicating the expected number of rows}

\item{partitioning}{A table partitioning is defined by a column name and a
vector of numeric values that are passed as \code{vec} argument to
\code{base::findInterval()}}

\item{x}{Object to test/coerce}

\item{fies}{A character vector of file names}
}
\description{
Used when importing a data source, a table specification defines column
data types, source file names, a mapping of source to imported column
names and optionally a number of expected rows and a partitioning scheme.
}
\keyword{internal}
