% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-class.R, R/config-utils.R
\name{new_tbl_cfg}
\alias{new_tbl_cfg}
\alias{is_tbl_cfg}
\alias{as_tbl_cfg}
\alias{as_tbl_cfg.tbl_cfg}
\alias{as_tbl_cfg.src_cfg}
\title{Table specification}
\usage{
new_tbl_cfg(
  table,
  src,
  files,
  cols,
  num_rows = NULL,
  partitioning = NULL,
  class_prefix = src
)

is_tbl_cfg(x)

as_tbl_cfg(x)

\method{as_tbl_cfg}{tbl_cfg}(x)

\method{as_tbl_cfg}{src_cfg}(x)
}
\arguments{
\item{table}{String valued table name}

\item{files}{A character vector of file names}

\item{cols}{List containing a list per column each holding string valued
entries \code{name} (column name as used by \code{ricu}), \code{col} (column name as used
in the raw data) and \code{spec} (name of \code{\link[readr:cols]{readr::cols()}} column specification).
Further entries will be passed as argument to the respective \code{readr} column
specification}

\item{partitioning}{A table partitioning is defined by a column name and a
vector of numeric values that are passed as \code{vec} argument to
\code{base::findInterval()}}

\item{x}{Object to test/coerce}

\item{nrow}{A count indicating the expected number of rows}
}
\description{
Used when importing a data source, a table specification defines column
data types, source file names, a mapping of source to imported column
names and optionally a number of expected rows and a partitioning scheme.
}
\keyword{internal}
