% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{new_tbl_spec}
\alias{new_tbl_spec}
\alias{is_tbl_spec}
\alias{as_tbl_spec}
\alias{as_tbl_spec.tbl_spec}
\alias{as_tbl_spec.list}
\alias{as_tbl_spec.src_cfg}
\title{Table specification}
\usage{
new_tbl_spec(table, files, cols, nrow, partitioning, src)

is_tbl_spec(x)

as_tbl_spec(x, ...)

\method{as_tbl_spec}{tbl_spec}(x, ...)

\method{as_tbl_spec}{list}(x, src, ...)

\method{as_tbl_spec}{src_cfg}(x, ...)
}
\arguments{
\item{table}{String valued table name}

\item{files}{A character vector of file names}

\item{cols}{List containing a list per column each holding string valued
entries \code{name} (column name as used by \code{ricu}), \code{col} (column name as used
in the raw data) and \code{spec} (name of \code{\link[readr:cols]{readr::cols()}} column specification).
Further entries will be passed as argument to the respective \code{readr} column
specification}

\item{nrow}{A count indicating the expected number of rows}

\item{partitioning}{A table partitioning is defined by a column name and a
vector of numeric values that are passed as \code{vec} argument to
\code{base::findInterval()}}

\item{x}{Object to test/coerce}
}
\description{
Used when importing a data source, a table specification defines column
data types, source file names, a mapping of source to imported column
names and optionally a number of expected rows and a partitioning scheme.
}
\keyword{internal}
