% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-utils.R
\name{expand}
\alias{expand}
\alias{has_no_gaps}
\alias{has_gaps}
\alias{fill_gaps}
\alias{slide}
\alias{slide_index}
\alias{hop}
\title{Time series utility functions}
\usage{
expand(
  x,
  min_col = "min",
  max_col = "max",
  id_vars = id_vars(x),
  step_size = time_step(x),
  new_index = index_var(x),
  keep = meta_vars(x)
)

has_no_gaps(x)

has_gaps(...)

fill_gaps(x, limits = NULL, ...)

slide(x, expr, before, after = hours(0L), ...)

slide_index(x, expr, index, before, after = hours(0L), ...)

hop(
  x,
  expr,
  windows,
  full_window = FALSE,
  lwr_col = "min_time",
  upr_col = "max_time",
  nomatch = NULL,
  eval_env = NULL
)
}
\arguments{
\item{x}{\code{ts_tbl} object to use}

\item{min_col, max_col}{Name of the columns that represent lower and upper
windows bounds}

\item{id_vars}{Names of the id columns}

\item{step_size}{Controls the step size used to interpolate between
\code{min_col} and \code{max_col}}

\item{new_index}{Name of the new index column}

\item{...}{Passed to \code{hop_quo()}}

\item{limits}{A table with columns for lower and upper window bounds}

\item{expr}{Expression (quoted for \verb{*_quo} and unquoted otherwise) to be
evaluated over each window}

\item{before, after}{Time span to look back/forward}

\item{index}{A vector of times around which windows are spanned (relative
to the index)}

\item{windows}{An \code{icu_tbl} defining the windows to span}

\item{full_window}{Logical flag controlling how the situation is handled
where the sliding window extends beyond available data}

\item{lwr_col, upr_col}{Names of columns (in \code{windows}) of lower/upper
window bounds}

\item{nomatch}{Forwarded to \code{\link[=data.table]{data.table::[()}}}

\item{eval_env}{Environment in which \code{expr} is substituted; \code{NULL} resolves
to the environment in which \code{expr} was created}
}
\description{
Time series utility functions
}
