% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.R
\name{cached_concept_loader}
\alias{cached_concept_loader}
\title{Internal utilities for data loading in vignettes}
\usage{
cached_concept_loader(prefix)
}
\arguments{
\item{prefix}{Object defining the row-subsetting}
}
\description{
Internal utilities for data loading in vignettes
}
\keyword{internal}
