% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-attach.R
\docType{data}
\name{attach_source}
\alias{attach_source}
\alias{attach_source.character}
\alias{attach_source.list}
\alias{attach_source.src_config}
\alias{setup_src_env}
\alias{setup_src_env.src_config}
\alias{data}
\alias{mimic}
\alias{mimic_demo}
\alias{eicu}
\alias{eicu_demo}
\alias{hirid}
\alias{get_src_env}
\alias{get_src_env.character}
\alias{get_src_env.default}
\alias{get_src_env.data_src}
\title{Data attach utilities}
\format{
An object of class \code{environment} of length 5.
}
\usage{
attach_source(x, ...)

\method{attach_source}{character}(x, file = NULL, ...)

\method{attach_source}{list}(x, ...)

\method{attach_source}{src_config}(x, dir = NULL, assign_env = .GlobalEnv, ...)

setup_src_env(x, env, dir = NULL)

\method{setup_src_env}{src_config}(x, env, dir = NULL)

data

get_src_env(x)

\method{get_src_env}{character}(x)

\method{get_src_env}{default}(x)

\method{get_src_env}{data_src}(x)
}
\arguments{
\item{x}{Data source to attach}

\item{...}{Forwarded to further calls to \code{attach_source()}}

\item{file}{File name string pointing to a non-default source configuration
file}

\item{dir}{Directory used to look for \code{\link[fst:fst]{fst::fst()}} files; \code{NULL} calls
\code{\link[=data_dir]{data_dir()}} using the source name as \code{subdir} argument}

\item{assign_env}{Environment in which the data source will become available}

\item{data_env, aux_env}{Environments where data tables and auxiliary tables
are assigned}
}
\description{
Attaching a data source sets up an environment containing \code{\link[prt:new_prt]{prt::new_prt()}}
files using \code{\link[base:delayedAssign]{base::delayedAssign()}}.
}
\keyword{datasets}
