% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-src.R, R/setup-attach.R
\docType{data}
\name{data}
\alias{data}
\alias{mimic}
\alias{mimic_demo}
\alias{eicu}
\alias{eicu_demo}
\alias{hirid}
\alias{new_src_tbl}
\alias{as_src_tbl}
\alias{as_src_tbl.src_tbl}
\alias{as_src_tbl.src_env}
\alias{as_src_tbl.character}
\alias{new_src_env}
\alias{as_src_env}
\alias{as_src_env.src_env}
\alias{as_src_env.character}
\alias{as_src_env.default}
\alias{as_src_env.src_tbl}
\alias{attach_src}
\alias{attach_src.src_cfg}
\alias{attach_src.character}
\alias{setup_src_env}
\alias{setup_src_env.src_cfg}
\title{Data attach utilities}
\format{
An object of class \code{environment} of length 5.
}
\usage{
data

new_src_tbl(files, col_cfg, tbl_cfg, prefix, src_env)

as_src_tbl(x, ...)

\method{as_src_tbl}{src_tbl}(x, ...)

\method{as_src_tbl}{src_env}(x, tbl, ...)

\method{as_src_tbl}{character}(x, env, ...)

new_src_env(x, env = new.env(parent = data_env()))

as_src_env(x)

\method{as_src_env}{src_env}(x)

\method{as_src_env}{character}(x)

\method{as_src_env}{default}(x)

\method{as_src_env}{src_tbl}(x)

attach_src(x, ...)

\method{attach_src}{src_cfg}(x, assign_env = .GlobalEnv, dir = src_data_dir(x), ...)

\method{attach_src}{character}(x, name = "data-sources", dir = NULL, assign_env = .GlobalEnv, ...)

setup_src_env(x, env, dir = NULL)

\method{setup_src_env}{src_cfg}(x, env, dir = src_data_dir(x))
}
\arguments{
\item{files}{File names of \code{fst} files that will be used to create a \code{prt}
object (see also \code{\link[prt:new_prt]{prt::new_prt()}})}

\item{col_cfg}{Coerced to \code{col_cfg} by calling \code{\link[=as_col_cfg]{as_col_cfg()}}}

\item{tbl_cfg}{Coerced to \code{tbl_cfg} by calling \code{\link[=as_tbl_cfg]{as_tbl_cfg()}}}

\item{prefix}{Character vector valued data source name(s) (used as class
prefix)}

\item{src_env}{The data source environment (as \code{src_env} object)}

\item{x}{Data source to attach}

\item{...}{Forwarded to further calls to \code{attach_src()}}

\item{tbl}{String valued table name}

\item{env}{Environment where data proxy objects are created}

\item{assign_env}{Environment in which the data source will become available}

\item{dir}{Directory used to look for \code{\link[fst:fst]{fst::fst()}} files; \code{NULL} calls
\code{\link[=data_dir]{data_dir()}} using the source name as \code{subdir} argument}

\item{name}{String valued name of a config file which will be looked up in
the default config directors}
}
\description{
Attaching a data source sets up an environment containing \code{\link[prt:new_prt]{prt::new_prt()}}
files using \code{\link[base:delayedAssign]{base::delayedAssign()}}.
}
\keyword{datasets}
