% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback-cncpt.R
\name{pafi}
\alias{pafi}
\alias{vent}
\alias{sed}
\alias{gcs}
\alias{urine24}
\alias{supp_o2}
\alias{avpu}
\title{Concept callback functions}
\usage{
pafi(
  ...,
  match_win = hours(2L),
  mode = c("match_vals", "extreme_vals", "fill_gaps"),
  fix_na_fio2 = TRUE,
  interval = NULL
)

vent(..., match_win = hours(6L), min_length = mins(10L), interval = NULL)

sed(..., interval = NULL)

gcs(
  ...,
  valid_win = hours(6L),
  set_sed_max = TRUE,
  set_na_max = TRUE,
  interval = NULL
)

urine24(..., min_win = hours(12L), interval = NULL)

supp_o2(..., interval = NULL)

avpu(..., interval = NULL)
}
\arguments{
\item{...}{Data input used for concept calculation}

\item{match_win}{Default ventilation window if no stop time can be
matched to a start time}

\item{mode}{Method for matching PaO₂ and FiO₂ values}

\item{fix_na_fio2}{Logical flag indicating whether to impute missing FiO₂
values with 21}

\item{interval}{Expected time series step size (determined from data if
\code{NULL})}

\item{min_length}{Minimal time span between a ventilation start and end
time}

\item{valid_win}{Maximal time window for which a GCS value is valid
if no newer measurement is available}

\item{set_sed_max}{Logical flag for considering sedation}

\item{set_na_max}{Logical flag controlling imputation of missing GCS values
with the respective maximum values}

\item{min_win}{Minimal time span required for calculation of urine/24h}
}
\description{
Owing to increased complexity and more diverse applications, recursive
concepts (class \code{\link[=new_cncpt]{rec_cncpt}}) may specify callback functions
to be called on corresponding data objects and perform post-processing
steps.
}
