% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{default_var}
\alias{default_var}
\alias{default_var.col_cfg}
\alias{default_var.id_cfg}
\alias{default_var.src_tbl}
\alias{src_name}
\alias{src_name.src_cfg}
\alias{src_name.id_cfg}
\alias{src_name.col_cfg}
\alias{src_name.tbl_cfg}
\alias{tbl_name}
\alias{tbl_name.col_cfg}
\alias{tbl_name.tbl_cfg}
\title{Get default columns}
\usage{
default_var(x, ...)

\method{default_var}{col_cfg}(x, type = "id_var", ...)

\method{default_var}{id_cfg}(x, ...)

\method{default_var}{src_tbl}(x, type = "id_var", ...)

src_name(x)

\method{src_name}{src_cfg}(x)

\method{src_name}{id_cfg}(x)

\method{src_name}{col_cfg}(x)

\method{src_name}{tbl_cfg}(x)

tbl_name(x)

\method{tbl_name}{col_cfg}(x)

\method{tbl_name}{tbl_cfg}(x)
}
\arguments{
\item{x}{Object used for dispatch}

\item{...}{Generic consistency}

\item{type}{The column type, e.g. \code{id}, \code{index}, \code{unit}, \code{value}}
}
\description{
For a table, query the default columns as specified by an \code{id_cfg}
(\code{\link[=new_id_cfg]{new_id_cfg()}}) or \code{col_cfg} (\code{\link[=new_col_cfg]{new_col_cfg()}}) object.
}
