% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{change_id}
\alias{change_id}
\alias{upgrade_id}
\alias{downgrade_id}
\alias{upgrade_id.ts_tbl}
\alias{upgrade_id.id_tbl}
\alias{downgrade_id.ts_tbl}
\alias{downgrade_id.id_tbl}
\title{Switch between id types}
\usage{
change_id(x, target_id, id_cfg, ...)

upgrade_id(x, target_id, id_cfg, cols = time_vars(x), ...)

downgrade_id(x, target_id, id_cfg, cols = time_vars(x), ...)

\method{upgrade_id}{ts_tbl}(x, target_id, id_cfg, cols = time_vars(x), ...)

\method{upgrade_id}{id_tbl}(x, target_id, id_cfg, cols = time_vars(x), ...)

\method{downgrade_id}{ts_tbl}(x, target_id, id_cfg, cols = time_vars(x), ...)

\method{downgrade_id}{id_tbl}(x, target_id, id_cfg, cols = time_vars(x), ...)
}
\arguments{
\item{x}{\code{icu_tbl} object for which to make the id change}

\item{target_id}{The destination id name}

\item{id_cfg}{An \code{id_cfg} object, specifying the id options}

\item{...}{Passed to \code{upgrade_id()}/\code{downgrade_id()}}

\item{cols}{Column names that require time-adjustment}
}
\description{
In ICU settings, multiple ID types may be in use which present an ordering
or nested structure, for example patient id, hospital stay id and ICU stay
id. This function allows for converting from one ID to another.
}
