% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{id_name}
\alias{id_name}
\alias{next_id}
\alias{prev_id}
\alias{change_id}
\alias{upgrade_id}
\alias{downgrade_id}
\alias{upgrade_id.ts_tbl}
\alias{upgrade_id.id_tbl}
\alias{downgrade_id.ts_tbl}
\alias{downgrade_id.id_tbl}
\title{Switch between id types}
\usage{
id_name(x, col = id(x))

next_id(x, n = 1L)

prev_id(x, n = 1L)

change_id(x, source, to, from = id_name(x), ...)

upgrade_id(
  x,
  source,
  to = next_id(x),
  from = id_name(x),
  cols = time_cols(x),
  ...
)

downgrade_id(
  x,
  source,
  to = prev_id(x),
  from = id_name(x),
  cols = time_cols(x),
  ...
)

\method{upgrade_id}{ts_tbl}(
  x,
  source,
  to = next_id(x),
  from = id_name(x),
  cols = time_cols(x),
  ...
)

\method{upgrade_id}{id_tbl}(
  x,
  source,
  to = next_id(x),
  from = id_name(x),
  cols = time_cols(x),
  interval = NULL,
  ...
)

\method{downgrade_id}{ts_tbl}(
  x,
  source,
  to = next_id(x),
  from = id_name(x),
  cols = time_cols(x),
  ...
)

\method{downgrade_id}{id_tbl}(
  x,
  source,
  to = next_id(x),
  from = id_name(x),
  cols = time_cols(x),
  interval = NULL,
  ...
)
}
\arguments{
\item{x}{\code{icu_tbl} object for which to make the id change}

\item{col}{Name of an id column}

\item{n}{Number of jumps in id ordering}

\item{source}{Name of the data source}

\item{to}{The destination id type}

\item{from}{The current id type}

\item{...}{Generic consistency}

\item{cols}{Column names that require time-adjustment}
}
\description{
In ICU settings, multiple ID types may be in use which present an ordering
or nested structure, for example patient id, hospital stay id and ICU stay
id. This function allows for converting from one ID to another.
}
