% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-env.R, R/setup-attach.R
\docType{data}
\name{data}
\alias{data}
\alias{aux}
\alias{attach_mimic}
\alias{attach_eicu}
\alias{attach_hirid}
\alias{attach_datasource}
\title{Data attach utilities}
\format{An object of class \code{environment} of length 5.}
\usage{
data

aux

attach_mimic(
  demo = FALSE,
  dir = mimic_data_dir(demo),
  config = mimic_config(demo),
  ...
)

attach_eicu(
  demo = FALSE,
  dir = eicu_data_dir(demo),
  config = eicu_config(demo),
  ...
)

attach_hirid(
  dir = data_dir("hirid", create = FALSE),
  config = get_config("hirid-setup"),
  ...
)

attach_datasource(dir, config, assign_env = .GlobalEnv)
}
\arguments{
\item{demo}{Logical flag indicating whether to import the respective demo
of full dataset. Only used for determining the correct default arguments
passed on to \code{attach_datasource()}.}

\item{dir}{Data directory where the imported \code{.fst} files are located (see
\code{\link[=import_datasource]{import_datasource()}} for further information).}

\item{config}{Import configuration specified as nested list (see
\code{\link[=import_datasource]{import_datasource()}} for further information).}
}
\description{
Data attach utilities
}
\keyword{datasets}
