% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R
\name{new_cncpt}
\alias{new_cncpt}
\alias{is_cncpt}
\alias{init_cncpt}
\alias{init_cncpt.num_cncpt}
\alias{init_cncpt.fct_cncpt}
\alias{init_cncpt.cncpt}
\alias{init_cncpt.rec_cncpt}
\alias{new_concept}
\alias{concept}
\alias{is_concept}
\title{Data concept}
\usage{
new_cncpt(
  name,
  items,
  description = NULL,
  aggregate = NULL,
  ...,
  class = "num_cncpt"
)

is_cncpt(x, ...)

init_cncpt(x, ...)

\method{init_cncpt}{num_cncpt}(x, unit = NULL, min = NULL, max = NULL, ...)

\method{init_cncpt}{fct_cncpt}(x, levels, ...)

\method{init_cncpt}{cncpt}(x, ...)

\method{init_cncpt}{rec_cncpt}(x, callback, target = c("ts_tbl", "id_tbl"), interval = NULL, ...)

new_concept(x)

concept(...)

is_concept(x)
}
\arguments{
\item{name}{The name of the concept}

\item{items}{Zero or more \code{itm} objects}

\item{description}{String-valued concept description}

\item{aggregate}{NULL or a string denoting a function used to aggregate per
id and if applicable per time step}

\item{...}{Further specification of the \code{cncpt} object (passed to
\code{\link[=init_cncpt]{init_cncpt()}})}

\item{class}{\code{NULL} or a string-valued sub-class name used for customizing
concept behavior}

\item{x}{Object to query/dispatch on}

\item{unit}{A string, specifying the measurement unit of the concept (can
be \code{NULL})}

\item{min, max}{Scalar valued; defines a range of plausible values for a
numeric concept}

\item{levels}{A vector of possible values a categorical concept may take on}

\item{callback}{Name of a function to be called on the returned data used
for data cleanup operations}

\item{target}{The target object yielded by loading}

\item{interval}{Time interval used for data loading; if NULL, the respective
interval passed as argument to \code{\link[=load_concepts]{load_concepts()}} is taken}
}
\description{
Clinical concepts are represented by \code{concept} objects.
}
