% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R, R/generics.R
\name{new_concept}
\alias{new_concept}
\alias{is_concept}
\alias{concept}
\alias{as_concept}
\title{Data concept}
\usage{
new_concept(name, items, unit = NULL)

is_concept(x)

concept(name, ..., unit = NULL)

as_concept(x, ...)
}
\arguments{
\item{name}{The name of the concept}

\item{items}{One or more \code{item} objects}

\item{unit}{A string, specifying the measurement unit of the concept (can
be \code{NULL})}

\item{x}{A concept object}

\item{...}{Forwarded to \code{\link[=new_item]{new_item()}}}
}
\description{
Clinical concepts are represented by \code{concept} objects.
}
