% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R, R/setup-attach.R
\name{new_concept}
\alias{new_concept}
\alias{is_concept}
\alias{read_dictionary}
\alias{src_name}
\title{Data concept}
\usage{
new_concept(
  name,
  items,
  unit = NULL,
  min = NULL,
  max = NULL,
  levels = NULL,
  target_class = "ts_tbl"
)

is_concept(x)

read_dictionary(src = NULL, name = "concept-dict", file = NULL, ...)

src_name(x)
}
\arguments{
\item{name}{Name of the dictionary to be read}

\item{items}{One or more \code{item} objects}

\item{unit}{A string, specifying the measurement unit of the concept (can
be \code{NULL})}

\item{min, max}{Scalar valued; defines a range of plausible values}

\item{x}{A concept object}

\item{file}{File name of the dictionary}
}
\description{
Clinical concepts are represented by \code{concept} objects.
}
