% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R
\name{new_cncpt}
\alias{new_cncpt}
\alias{is_cncpt}
\alias{init_cncpt}
\alias{init_cncpt.num_cncpt}
\alias{init_cncpt.fct_cncpt}
\alias{init_cncpt.cncpt}
\alias{init_cncpt.rec_cncpt}
\alias{new_concept}
\alias{concept}
\alias{is_concept}
\title{Data items}
\usage{
new_cncpt(
  name,
  items,
  description = NA_character_,
  category = NA_character_,
  aggregate = NULL,
  ...,
  class = "num_cncpt"
)

is_cncpt(x, ...)

init_cncpt(x, ...)

\method{init_cncpt}{num_cncpt}(x, unit = NULL, min = NULL, max = NULL, ...)

\method{init_cncpt}{fct_cncpt}(x, levels, ...)

\method{init_cncpt}{cncpt}(x, ...)

\method{init_cncpt}{rec_cncpt}(x, callback, target = c("ts_tbl", "id_tbl"), interval = NULL, ...)

new_concept(x)

concept(...)

is_concept(x)
}
\arguments{
\item{name}{The name of the concept}

\item{items}{Zero or more \code{itm} objects}

\item{description}{String-valued concept description}

\item{category}{String-valued category}

\item{aggregate}{NULL or a string denoting a function used to aggregate per
id and if applicable per time step}

\item{...}{Further specification of the \code{cncpt} object (passed to
\code{\link[=init_cncpt]{init_cncpt()}})}

\item{class}{\code{NULL} or a string-valued sub-class name used for customizing
concept behavior}

\item{x}{Object to query/dispatch on}

\item{unit}{A string, specifying the measurement unit of the concept (can
be \code{NULL})}

\item{min, max}{Scalar valued; defines a range of plausible values for a
numeric concept}

\item{levels}{A vector of possible values a categorical concept may take on}

\item{callback}{Name of a function to be called on the returned data used
for data cleanup operations}

\item{target}{The target object yielded by loading}

\item{interval}{Time interval used for data loading; if NULL, the respective
interval passed as argument to \code{\link[=load_concepts]{load_concepts()}} is taken}
}
\description{
Concept objects are used in \code{ricu} as a way to specify how a clinical
concept, such as heart rate can be loaded from a data source and are mainly
consumed by \code{\link[=load_concepts]{load_concepts()}}. Several functions are avialable for
constructing \code{concept} (and related auxillary) objects either from code or
by parsing a JSON formatted concept dictionary using \code{\link[=load_dicionary]{load_dicionary()}}.
}
\details{
:
In order to allow for a large degree of flexibility (and extensibility),
which is much needed owing to considerable heterogeneity presented by
different data sources, several nested S3 classes are involved in
representing a concept. An outline of this hierarchy can be described as
\itemize{
\item \code{concept}: contains many \code{cncpt} objects (of potentially differing
sub-types), each comprising of some meta-data and an \code{item} object
\item \code{item}: contains many \code{itm} objects (of potentially differing
sub-types), each encoding how to retrieve a data item.
}

The design choice for wrapping a vector of \code{cncpt} objects with a container
class \code{concept} is motivated by the requirement of having several different
sub-types of \code{cncpt} objects (all inheriting from the parent type \code{cncpt}),
while retaining control over how this homogeneous w.r.t. parent type, but
heterogeneous w.r.t. sub-type vector of objects behaves in terms of S3
generic functions.
}
