% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R
\name{item}
\alias{item}
\alias{as_item}
\alias{n_tick}
\alias{n_tick.itm}
\alias{n_tick.item}
\alias{new_cncpt}
\alias{is_cncpt}
\alias{init_cncpt}
\alias{init_cncpt.num_cncpt}
\alias{init_cncpt.fct_cncpt}
\alias{init_cncpt.cncpt}
\alias{init_cncpt.rec_cncpt}
\alias{src_name.cncpt}
\alias{n_tick.cncpt}
\alias{new_concept}
\alias{concept}
\alias{is_concept}
\alias{n_tick.concept}
\alias{load_dictionary}
\title{Data concept}
\usage{
item(...)

as_item(x)

n_tick(x)

\method{n_tick}{itm}(x)

\method{n_tick}{item}(x)

new_cncpt(
  name,
  items,
  description = NULL,
  aggregate = NULL,
  ...,
  class = "num_cncpt"
)

is_cncpt(x, ...)

init_cncpt(x, ...)

\method{init_cncpt}{num_cncpt}(x, unit = NULL, min = NULL, max = NULL, ...)

\method{init_cncpt}{fct_cncpt}(x, levels, ...)

\method{init_cncpt}{cncpt}(x, ...)

\method{init_cncpt}{rec_cncpt}(x, callback, target = c("ts_tbl", "id_tbl"), interval = NULL, ...)

\method{src_name}{cncpt}(x)

\method{n_tick}{cncpt}(x)

new_concept(x)

concept(...)

is_concept(x)

\method{n_tick}{concept}(x)

load_dictionary(
  src = NULL,
  concepts = NULL,
  name = "concept-dict",
  file = NULL
)
}
\arguments{
\item{...}{Further specification of the \code{cncpt} object (passed to
\code{\link[=init_cncpt]{init_cncpt()}})}

\item{x}{Object to query/dispatch on}

\item{name}{Name of the dictionary to be read}

\item{items}{Zero or more \code{itm} objects}

\item{description}{String-valued concept description}

\item{aggregate}{NULL or a string denoting a function used to aggregate per
id and if applicable per time step}

\item{class}{\code{NULL} or a string-valued sub-class name used for customizing
concept behavior}

\item{unit}{A string, specifying the measurement unit of the concept (can
be \code{NULL})}

\item{min, max}{Scalar valued; defines a range of plausible values for a
numeric concept}

\item{levels}{A vector of possible values a categorical concept may take on}

\item{callback}{Name of a function to be called on the returned data used
for data cleanup operations}

\item{target}{The target object yielded by loading}

\item{interval}{Time interval used for data loading; if NULL, the respective
interval passed as argument to \code{\link[=load_concepts]{load_concepts()}} is taken}

\item{src}{\code{NULL} or the name of a data source}

\item{concepts}{A character vector used to subset the concept dictionary or
\code{NULL} indicating no subsetting}

\item{file}{File name of the dictionary}
}
\description{
Clinical concepts are represented by \code{concept} objects.
}
