% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R, R/setup-attach.R
\name{item}
\alias{item}
\alias{as_item}
\alias{new_cncpt}
\alias{is_cncpt}
\alias{init_cncpt}
\alias{init_cncpt.num_cncpt}
\alias{init_cncpt.fct_cncpt}
\alias{init_cncpt.cncpt}
\alias{src_name.cncpt}
\alias{new_concept}
\alias{concept}
\alias{is_concept}
\alias{read_dictionary}
\alias{src_name}
\title{Data concept}
\usage{
item(...)

as_item(x)

new_cncpt(name, items, ..., class = "num_cncpt")

is_cncpt(x, ...)

init_cncpt(x, ...)

\method{init_cncpt}{num_cncpt}(x, unit = NULL, min = NULL, max = NULL, ...)

\method{init_cncpt}{fct_cncpt}(x, levels, ...)

\method{init_cncpt}{cncpt}(x, ...)

\method{src_name}{cncpt}(x)

new_concept(x)

concept(...)

is_concept(x)

read_dictionary(
  src = NULL,
  concepts = NULL,
  name = "concept-dict",
  file = NULL,
  ...
)

src_name(x)
}
\arguments{
\item{...}{Further specification of the \code{cncpt} object (passed to
\code{\link[=init_cncpt]{init_cncpt()}})}

\item{x}{Object to query/dispatch on}

\item{name}{Name of the dictionary to be read}

\item{items}{Zero or more \code{itm} objects}

\item{class}{\code{NULL} or a string-valued sub-class name used for customizing
concept behavior}

\item{unit}{A string, specifying the measurement unit of the concept (can
be \code{NULL})}

\item{min, max}{Scalar valued; defines a range of plausible values for a
numeric concept}

\item{levels}{A vector of possible values a categorical concept may take on}

\item{src}{\code{NULL} or the name of a data source}

\item{concepts}{A character vector used to subset the concept dictionary or
\code{NULL} indicating no subsetting}

\item{file}{File name of the dictionary}
}
\description{
Clinical concepts are represented by \code{concept} objects.
}
