% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R
\name{new_itm}
\alias{new_itm}
\alias{is_itm}
\alias{init_itm}
\alias{init_itm.sel_itm}
\alias{init_itm.col_itm}
\alias{init_itm.rgx_itm}
\alias{init_itm.los_itm}
\alias{init_itm.itm}
\alias{new_item}
\alias{is_item}
\title{Data item}
\usage{
new_itm(src, ..., target = c("ts_tbl", "id_tbl"), class = "sel_itm")

is_itm(x)

init_itm(x, ...)

\method{init_itm}{sel_itm}(
  x,
  table,
  sub_col,
  ids,
  itm_cols = NULL,
  index_col = NULL,
  callback = "identity_callback",
  ...
)

\method{init_itm}{col_itm}(
  x,
  table,
  itm_cols = NULL,
  index_col = NULL,
  callback = "identity_callback",
  ...
)

\method{init_itm}{rgx_itm}(
  x,
  table,
  sub_col,
  regex,
  itm_cols = NULL,
  index_col = NULL,
  callback = "identity_callback",
  ...
)

\method{init_itm}{los_itm}(x, win_type, ...)

\method{init_itm}{itm}(x, ...)

new_item(x)

is_item(x)
}
\arguments{
\item{src}{The data source name}

\item{...}{Further specification of the \code{itm} object (passed to
\code{\link[=init_itm]{init_itm()}})}

\item{target}{The target object yielded by loading}

\item{class}{Sub class for customizing \code{itm} behavior}

\item{x}{Object to query/dispatch on}

\item{table}{Name of the table containing the data}

\item{sub_col}{Column name used for subsetting}

\item{ids}{Vector of ids used to subset table rows. If \code{NULL}, all rows are
considered corresponding to the data item}

\item{itm_cols}{Columns returned as \code{\link[=data_vars]{data_vars()}}}

\item{index_col}{Column used as index}

\item{callback}{Name of a function to be called on the returned data used
for data cleanup operations}

\item{regex}{String-valued regular expression which will be evaluated by
\code{\link[base:grepl]{base::grepl()}} with \code{ignore.case = TRUE}}

\item{win_type}{Passed to \code{\link[=stay_windows]{stay_windows()}}}
}
\description{
In order to specify the location of data items in data sets, \code{item} objects
are used.
}
