% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R
\name{new_item}
\alias{new_item}
\alias{item}
\alias{is_item}
\alias{as_item}
\alias{src_name}
\title{Data item}
\usage{
new_item(
  concept,
  source,
  table = NULL,
  column = NULL,
  ids = NULL,
  regex = FALSE,
  callback = NULL,
  load_fun = "load_default",
  ...
)

item(...)

is_item(x)

as_item(x)

src_name(x)
}
\arguments{
\item{concept}{The name of the associated clinical concept}

\item{source}{The data source name}

\item{table}{Name of the table containing the data}

\item{column}{Column name}

\item{ids}{Vector of ids used to subset table rows. If \code{NULL}, all rows are
considered corresponding to the data item}

\item{regex}{Logical flag indicating whether the vector of id should be
interpreted as regular expressions}

\item{callback}{Name of a function to be called on the returned data used
for data cleanup operations}

\item{load_fun}{Custom function used for overriding default loading of data}

\item{...}{Further name/string pairs specifying additional columns needed to
define the data item}

\item{x}{A data item object}
}
\description{
In order to specify the location of data items in data sets, \code{item} objects
are used.
}
