% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-utils.R
\name{new_item}
\alias{new_item}
\alias{is_item}
\alias{validate_itm}
\alias{validate_itm.sel_itm}
\alias{validate_itm.col_itm}
\alias{validate_itm.rgx_itm}
\alias{validate_itm.los_itm}
\alias{validate_itm.itm}
\alias{as_item}
\alias{as_item.item}
\title{Data item}
\usage{
new_item(src, itm, class = "sel_itm")

is_item(x)

validate_itm(x, src)

\method{validate_itm}{sel_itm}(x, src)

\method{validate_itm}{col_itm}(x, src)

\method{validate_itm}{rgx_itm}(x, src)

\method{validate_itm}{los_itm}(x, src)

\method{validate_itm}{itm}(x, src)

as_item(x, ...)

\method{as_item}{item}(x, ...)
}
\arguments{
\item{src}{The data source name}

\item{x}{A data item object}

\item{...}{Further name/string pairs specifying additional columns needed to
define the data item}

\item{table}{Name of the table containing the data}

\item{column}{Column name}

\item{ids}{Vector of ids used to subset table rows. If \code{NULL}, all rows are
considered corresponding to the data item}

\item{regex}{Logical flag indicating whether the vector of id should be
interpreted as regular expressions}

\item{callback}{Name of a function to be called on the returned data used
for data cleanup operations}

\item{load_fun}{Custom function used for overriding default loading of data}
}
\description{
In order to specify the location of data items in data sets, \code{item} objects
are used.
}
