% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{id_origin}
\alias{id_origin}
\alias{id_orig_helper}
\alias{id_orig_helper.src_env}
\alias{id_windows}
\alias{id_win_helper}
\alias{id_win_helper.mimic_env}
\alias{id_win_helper.eicu_env}
\alias{id_win_helper.hirid_env}
\alias{id_map}
\alias{id_map_helper}
\alias{id_map_helper.src_env}
\title{Data loading utilities}
\usage{
id_origin(x, id, origin_name = NULL, copy = TRUE)

id_orig_helper(x, id)

\method{id_orig_helper}{src_env}(x, id)

id_windows(x, copy = TRUE)

id_win_helper(x)

\method{id_win_helper}{mimic_env}(x)

\method{id_win_helper}{eicu_env}(x)

\method{id_win_helper}{hirid_env}(x)

id_map(x, id_var, win_var, in_time = NULL, out_time = NULL)

id_map_helper(x, id_var, win_var)

\method{id_map_helper}{src_env}(x, id_var, win_var)
}
\arguments{
\item{x}{Object identify the ID system (passed to \code{\link[=as_src_env]{as_src_env()}})}

\item{id}{ID name for which to return origin times}

\item{origin_name}{String-valued name which will be used to label the origin
column}

\item{copy}{Logical flag indicating whether to return a copy of the memoized
\verb{0data.table} for safety}

\item{id_var}{Type of ID all returned times are relative to}

\item{win_var}{Type of ID for which the in/out times is returned}

\item{in_time, out_time}{column names of the returned in/out times}
}
\description{
Two important tools for smoothing out differences among used datasets are
\code{id_origin()} which returns origin times for a given ID and \code{id_windows()}
which defines stay windows and relationships between ID systems.
}
\keyword{internal}
