% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{id_orig}
\alias{id_orig}
\alias{id_orig.id_cfg}
\alias{id_orig.default}
\alias{id_origin}
\alias{id_win_min}
\alias{id_win_min.mimic_ids}
\alias{id_win_min.eicu_ids}
\alias{id_win_min.hirid_ids}
\alias{id_win_min.default}
\alias{id_windows}
\title{Data loading utilities}
\usage{
id_orig(x, ...)

\method{id_orig}{id_cfg}(x, id, origin_name = NULL, env = as_src_env(x), ...)

\method{id_orig}{default}(x, id, env = as_src_env(x), ...)

id_origin(x, ...)

id_win_min(x, ...)

\method{id_win_min}{mimic_ids}(x, env = as_src_env(x), ...)

\method{id_win_min}{eicu_ids}(x, env = as_src_env(x), ...)

\method{id_win_min}{hirid_ids}(x, env = as_src_env(x), ...)

\method{id_win_min}{default}(x, env = as_src_env(x), ...)

id_windows(x, ...)
}
\arguments{
\item{x}{Object to dispatch on}

\item{...}{Generic consistency}

\item{id}{ID name for which to return origin times}

\item{origin_name}{String-valued name which will be used to label the origin
column}

\item{env}{Data environment in which to look for tables to construct helper
tables}
}
\description{
Two important tools for smoothing out differences among used datasets are
\code{id_origin()} which returns origin times for a given ID and \code{id_windows()}
which defines stay windows and relationships between ID systems.
}
