% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/zzz-assertions.R
\name{secs}
\alias{secs}
\alias{mins}
\alias{hours}
\alias{days}
\alias{weeks}
\alias{is_time}
\alias{is_time_vec}
\title{Difftime utilities}
\usage{
secs(x)

mins(x)

hours(x)

days(x)

weeks(x)

is_time(x, allow_neg = TRUE)

is_time_vec(x, allow_neg = TRUE)
}
\arguments{
\item{x}{Object to test/coerce}
}
\description{
Utility functions for working with \code{\link[base:difftime]{base::difftime()}} objects.
}
