% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{secs}
\alias{secs}
\alias{mins}
\alias{hours}
\alias{days}
\alias{weeks}
\title{Difftime utilities}
\usage{
secs(x)

mins(x)

hours(x)

days(x)

weeks(x)
}
\arguments{
\item{x}{Object to test/coerce}
}
\description{
Utility functions for working with \code{\link[base:difftime]{base::difftime()}} objects.
}
