% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ts.R
\name{secs}
\alias{secs}
\alias{mins}
\alias{hours}
\alias{days}
\alias{weeks}
\title{Difftime utilities}
\usage{
secs(x)

mins(x)

hours(x)

days(x)

weeks(x)
}
\arguments{
\item{x}{Numeric vector to coerce to \code{\link[base:difftime]{base::difftime()}}}
}
\description{
As \code{\link[base:difftime]{base::difftime()}} vectors are used throughout \code{ricu}, a set of wrapper
functions are exported for convenience of instantiation \code{\link[base:difftime]{base::difftime()}}
vectors with given time units.
}
\examples{
hours(1L)
mins(NA_real_)
secs(1:10)
hours(numeric(0L))

}
