% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_dir}
\alias{data_dir}
\alias{get_config}
\title{File system utilities}
\usage{
data_dir(subdir = NULL)

get_config(name, dir = NULL, ...)
}
\arguments{
\item{subdir}{A string specifying a directory that will be made sure to
exist below the data directory.}

\item{name}{File name of the configuration file (without file ending).}

\item{dir}{(Optional) directory name where to look for the specified file.}

\item{...}{Passed to \link[jsonlite:read_json]{jsonlite::read_json}.}
}
\description{
Determine the location where to place data meant to persist between
individual sessions.
}
\details{
For data, the default location depends on the operating system as

| \strong{Platform} | \strong{Location}                          |
| ------------ | ------------------------------------- |
| Linux        | \code{~/.local/share/sepsr}                |
| macOS        | \code{~/Library/Application Support/sepsr} |
| Windows      | \code{\%LOCALAPPDATA\%/sepsr}                |

If the default storage directory does not exists, it will only be created
upon user consent (requiring an interactive session).

The environment variable \code{SEPSR_DATA_PATH} can be used to overwrite the
default location. If desired, this variable can be set in an R startup file
to make it apply to all R sessions. For example, it could be set within:
\itemize{
\item A project-local \code{.Renviron};
\item The user-level \code{.Renviron};
\item A file at \code{$(R RHOME)/etc/Renviron.site}.
}

Any directory specified as environment variable will recursively be created.
}
\examples{
Sys.setenv(SEPSR_DATA_PATH = tempdir())
identical(data_dir(), tempdir())

dir.exists(file.path(tempdir(), "some_subdir"))
some_subdir <- data_dir("some_subdir")
dir.exists(some_subdir)

cfg <- get_config("eicu-demo")
identical(
  cfg,
  get_config("eicu-demo",
             system.file("extdata", "config", package = "sepsr"))
)

}
