% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{get_src_config}
\alias{get_src_config}
\alias{get_src_config.src_config}
\alias{get_src_config.character}
\alias{get_src_config.default}
\title{Load configuration for a data source}
\usage{
get_src_config(x, ...)

\method{get_src_config}{src_config}(x, ...)

\method{get_src_config}{character}(x, name = "data-sources", file = NULL, ...)

\method{get_src_config}{default}(x, ...)
}
\arguments{
\item{x}{Object used to determine the data source}

\item{...}{Generic consistency/passed on to further \code{get_src_config()} calls}

\item{name}{String valued name of a config file which will be looked up in
the default config directors}

\item{file}{Full file name to load}
}
\description{
For a data source to become available to ricu, a JSON base configuration
file is required which is parsed into a \code{src_config} object.
}
