% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-id.R
\name{id_tbl}
\alias{id_tbl}
\alias{as_id_tbl.id_tbl}
\alias{as_id_tbl.default}
\alias{as_ts_tbl.id_tbl}
\alias{is_id_tbl}
\title{Tabular ICU data}
\usage{
id_tbl(..., id, id_opts = NULL)

\method{as_id_tbl}{id_tbl}(x, ...)

\method{as_id_tbl}{default}(x, id, id_opts = NULL, ...)

\method{as_ts_tbl}{id_tbl}(x, index = NULL, interval = hours(1L), ...)

is_id_tbl(x)
}
\arguments{
\item{...}{forwarded to \code{\link[data.table:data.table]{data.table::data.table()}} or generic consistency}

\item{id}{String valued column name used as \code{id} column}

\item{id_opts}{\code{NULL} or a named character vector indicating options to
which ids may be converted \code{\link[=see change_id]{see change_id()}}}

\item{x}{Object to test/coerce}

\item{index}{String valued column name used as index}

\item{interval}{Scalar \code{difftime} object representing the minimal time
difference between rows}
}
\description{
Two objects central to \code{ricu}, \code{id_tbl} and \code{ts_tbl} (both inheriting
from \code{icu_tbl}) are used to represent tabular data.
}
