% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{new_id_cfg}
\alias{new_id_cfg}
\alias{is_id_cfg}
\alias{as_id_cfg}
\alias{as_id_cfg.list}
\alias{as_id_cfg.id_cfg}
\alias{as_id_cfg.src_cfg}
\alias{as_id_cfg.data_src}
\title{ID columns}
\usage{
new_id_cfg(name, id, pos, start, end, table, src)

is_id_cfg(x)

as_id_cfg(x, ...)

\method{as_id_cfg}{list}(x, src, ...)

\method{as_id_cfg}{id_cfg}(x, ...)

\method{as_id_cfg}{src_cfg}(x, ...)

\method{as_id_cfg}{data_src}(x, ...)
}
\arguments{
\item{x}{Object to test/coerce}

\item{cfg}{List containing string valued column names}
}
\description{
A set of id columns corresponding to \code{patient}, \code{hadm} (hospital stay) and
\code{icustay} (ICU stay) can be represented by an \code{id_cfg} object.
}
\keyword{internal}
