% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{new_id_cols}
\alias{new_id_cols}
\alias{is_id_cols}
\alias{as_id_cols}
\alias{as_id_cols.id_cols}
\alias{as_id_cols.list}
\alias{as_id_cols.src_config}
\title{ID columns}
\usage{
new_id_cols(cfg)

is_id_cols(x)

as_id_cols(x)

\method{as_id_cols}{id_cols}(x)

\method{as_id_cols}{list}(x)

\method{as_id_cols}{src_config}(x)
}
\arguments{
\item{cfg}{List containing string valued column names}

\item{x}{Object to test/coerce}
}
\description{
A set of id columns corresponding to \code{patient}, \code{hadm} (hospital stay) and
\code{icustay} (ICU stay) can be represented by an \code{id_cols} object.
}
\keyword{internal}
