% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-class.R
\name{id_tbl}
\alias{id_tbl}
\alias{is_id_tbl}
\alias{as_id_tbl}
\alias{as_id_tbl.ts_tbl}
\alias{as_id_tbl.id_tbl}
\alias{as_id_tbl.data.table}
\alias{as_id_tbl.default}
\alias{ts_tbl}
\alias{is_ts_tbl}
\alias{as_ts_tbl}
\alias{as_ts_tbl.ts_tbl}
\alias{as_ts_tbl.id_tbl}
\alias{as_ts_tbl.data.table}
\alias{as_ts_tbl.default}
\alias{unclass_tbl}
\alias{unclass_tbl.data.frame}
\alias{unclass_tbl.ts_tbl}
\alias{unclass_tbl.id_tbl}
\alias{reclass_tbl}
\alias{reclass_tbl.NULL}
\alias{reclass_tbl.id_tbl}
\alias{reclass_tbl.ts_tbl}
\alias{as_ptype}
\alias{as_ptype.id_tbl}
\alias{as_ptype.ts_tbl}
\alias{validate_tbl}
\alias{validate_tbl.id_tbl}
\alias{validate_tbl.ts_tbl}
\alias{validate_tbl.data.table}
\title{Tabular ICU data}
\usage{
id_tbl(..., id_vars = 1L)

is_id_tbl(x)

as_id_tbl(x, id_vars = NULL, by_ref = FALSE)

\method{as_id_tbl}{ts_tbl}(x, id_vars = NULL, by_ref = FALSE)

\method{as_id_tbl}{id_tbl}(x, id_vars = NULL, by_ref = FALSE)

\method{as_id_tbl}{data.table}(x, id_vars = NULL, by_ref = FALSE)

\method{as_id_tbl}{default}(x, id_vars = NULL, by_ref = FALSE)

ts_tbl(..., id_vars = 1L, index_var = NULL, interval = NULL)

is_ts_tbl(x)

as_ts_tbl(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

\method{as_ts_tbl}{ts_tbl}(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

\method{as_ts_tbl}{id_tbl}(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

\method{as_ts_tbl}{data.table}(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

\method{as_ts_tbl}{default}(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

unclass_tbl(x)

\method{unclass_tbl}{data.frame}(x)

\method{unclass_tbl}{ts_tbl}(x)

\method{unclass_tbl}{id_tbl}(x)

reclass_tbl(x, template, stop_on_fail = TRUE)

\method{reclass_tbl}{`NULL`}(x, template, stop_on_fail = TRUE)

\method{reclass_tbl}{id_tbl}(x, template, stop_on_fail = TRUE)

\method{reclass_tbl}{ts_tbl}(x, template, stop_on_fail = TRUE)

as_ptype(x)

\method{as_ptype}{id_tbl}(x)

\method{as_ptype}{ts_tbl}(x)

validate_tbl(x)

\method{validate_tbl}{id_tbl}(x)

\method{validate_tbl}{ts_tbl}(x)

\method{validate_tbl}{data.table}(x)
}
\arguments{
\item{...}{forwarded to \code{\link[data.table:data.table]{data.table::data.table()}} or generic consistency}

\item{id_vars}{Column name(s) to be used as \code{id} column(s)}

\item{x}{Object to query/operate on}

\item{by_ref}{Logical flag indicating whether to perform the operation by
reference}

\item{index_var}{Column name of the index column}

\item{interval}{Time series interval length specified as scalar-valued
\code{difftime} object}

\item{template}{Object after which to model the object in question}

\item{stop_on_fail}{Logical flag indicating whether to consider failed
object validation as error}
}
\description{
Two objects central to \code{ricu}, \code{id_tbl} and \code{ts_tbl} (both inheriting
from \code{id_tbl}) are used to represent tabular data.
}
