% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-class.R
\name{id_tbl}
\alias{id_tbl}
\alias{is_id_tbl}
\alias{as_id_tbl}
\alias{as_id_tbl.ts_tbl}
\alias{as_id_tbl.id_tbl}
\alias{as_id_tbl.data.table}
\alias{as_id_tbl.default}
\alias{ts_tbl}
\alias{is_ts_tbl}
\alias{as_ts_tbl}
\alias{as_ts_tbl.ts_tbl}
\alias{as_ts_tbl.id_tbl}
\alias{as_ts_tbl.data.table}
\alias{as_ts_tbl.default}
\alias{unclass_tbl}
\alias{unclass_tbl.data.frame}
\alias{unclass_tbl.ts_tbl}
\alias{unclass_tbl.id_tbl}
\alias{reclass_tbl}
\alias{reclass_tbl.NULL}
\alias{reclass_tbl.id_tbl}
\alias{reclass_tbl.ts_tbl}
\alias{as_ptype}
\alias{as_ptype.id_tbl}
\alias{as_ptype.ts_tbl}
\alias{validate_tbl}
\alias{validate_tbl.id_tbl}
\alias{validate_tbl.ts_tbl}
\alias{validate_tbl.data.table}
\title{Tabular ICU data}
\usage{
id_tbl(..., id_vars = 1L)

is_id_tbl(x)

as_id_tbl(x, id_vars = NULL, by_ref = FALSE)

\method{as_id_tbl}{ts_tbl}(x, id_vars = NULL, by_ref = FALSE)

\method{as_id_tbl}{id_tbl}(x, id_vars = NULL, by_ref = FALSE)

\method{as_id_tbl}{data.table}(x, id_vars = NULL, by_ref = FALSE)

\method{as_id_tbl}{default}(x, id_vars = NULL, by_ref = FALSE)

ts_tbl(..., id_vars = 1L, index_var = NULL, interval = NULL)

is_ts_tbl(x)

as_ts_tbl(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

\method{as_ts_tbl}{ts_tbl}(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

\method{as_ts_tbl}{id_tbl}(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

\method{as_ts_tbl}{data.table}(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

\method{as_ts_tbl}{default}(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

unclass_tbl(x)

\method{unclass_tbl}{data.frame}(x)

\method{unclass_tbl}{ts_tbl}(x)

\method{unclass_tbl}{id_tbl}(x)

reclass_tbl(x, template, stop_on_fail = TRUE)

\method{reclass_tbl}{`NULL`}(x, template, stop_on_fail = TRUE)

\method{reclass_tbl}{id_tbl}(x, template, stop_on_fail = TRUE)

\method{reclass_tbl}{ts_tbl}(x, template, stop_on_fail = TRUE)

as_ptype(x)

\method{as_ptype}{id_tbl}(x)

\method{as_ptype}{ts_tbl}(x)

validate_tbl(x)

\method{validate_tbl}{id_tbl}(x)

\method{validate_tbl}{ts_tbl}(x)

\method{validate_tbl}{data.table}(x)
}
\arguments{
\item{...}{forwarded to \code{\link[data.table:data.table]{data.table::data.table()}} or generic consistency}

\item{id_vars}{Column name(s) to be used as \code{id} column(s)}

\item{x}{Object to query/operate on}

\item{by_ref}{Logical flag indicating whether to perform the operation by
reference}

\item{index_var}{Column name of the index column}

\item{interval}{Time series interval length specified as scalar-valued
\code{difftime} object}

\item{template}{Object after which to model the object in question}

\item{stop_on_fail}{Logical flag indicating whether to consider failed
object validation as error}
}
\description{
In order to simplify handling or tabular ICU data, \code{ricu} provides two
S3 classes, \code{id_tbl} and \code{ts_tbl}. The two classes essentially
consist of a \code{data.table} object, alongside some meta data and S3 dispatch
is used to enable more natural behavior for some data manipulation tasks.
For example, when merging two tables, a default for the \code{by} argument can
be chosen more sensibly if columns representing patient ID and timestamp
information can be identified.
}
\details{
The two classes are designed for two often encoutered data scenatios:
\itemize{
\item \code{id_tbl} objects can be used to represent static (with repspect to
relevant time scales) patient data such as patient age and such an object
is simply a \code{data.table} combined with a non-zero length character vector
valued attribute marking the columns tracking patient ID information
(\link[=id_vars]{id_vars}). All further columns are considered as
\link[=data_vars]{data_vars}.
\item \code{ts_tbl} objects are used for grouped time series data. A \code{data.table}
object again is augmented by attributes, including a non-zero length
character vector identifying patient ID columns (\link[=id_vars]{id_vars}),
a string, tracking the column holding time-stamps
(\link[=index_var]{index_var}) and a scalar \code{difftime} object determining
the time-series step size \link[=interval]{interval}. Again, all further
columns are treated as \link[=data_vars]{data_vars}.
Owing to the nested structure of required meta data, \code{ts_tbl} inherits from
ìd_tbl\verb{. Furthermore, both classes inherit from }data.table\verb{. As such, }data.table\verb{[reference semantics][data.table::reference] are available for some operations, indicated by presence of a}by_ref\verb{argument. At default, value,}by_ref\verb{is disabled as this is inline with base R behavior at the cost of potentially incurring unnecessary data copies. Some care has to be taken when passing}by_ref = TRUE` and enabling by reference operations as
this can have side effects (see examples).
}
}
