% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-load.R, R/concept-utils.R
\name{load_cncpt}
\alias{load_cncpt}
\alias{load_itm}
\alias{prepare_query}
\alias{add_unit_var}
\title{Internal utilities for \code{item}/\code{concept} objects}
\usage{
load_cncpt(x, verbose, ...)

load_itm(x, patient_ids, id_type, interval)

prepare_query(x)

add_unit_var(x)
}
\arguments{
\item{x}{Object defining the row-subsetting}

\item{verbose}{Logical flag for muting informational output}

\item{...}{Passed to downstream methods}

\item{patient_ids}{Optional vector of patient ids to subset the fetched data
with}

\item{id_type}{String specifying the patient id type to return}

\item{interval}{The time interval used to discretize time stamps with,
specified as \code{\link[base:difftime]{base::difftime()}} object}
}
\description{
Internal utilities for \code{item}/\code{concept} objects
}
\keyword{internal}
