% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-si.R
\name{si_data}
\alias{si_data}
\alias{si_windows}
\alias{si}
\title{Suspicion of infection label}
\usage{
si_data(
  source,
  abx_count_win = hours(24L),
  abx_min_count = 1L,
  positive_cultures = FALSE,
  dictionary = read_dictionary("concept-dict"),
  ...
)

si_windows(
  tbl,
  si_mode = c("and", "or"),
  abx_win = hours(24L),
  samp_win = hours(72L),
  si_lwr = hours(48L),
  si_upr = hours(24L)
)

si(source, ...)
}
\arguments{
\item{source}{String valued name of data source}

\item{abx_count_win}{Time span during which to apply the \code{abx_min_count}
criterion}

\item{abx_min_count}{Minimal number of antibiotic administrations}

\item{positive_cultures}{Logical flag indicating whether to require
cultures to be positive}

\item{dictionary}{\code{dictionary} object to use for concept lookup}

\item{...}{Passed to \code{\link[=load_concepts]{load_concepts()}}}

\item{tbl}{\code{ts_tbl} object to use for label computations}

\item{si_mode}{Switch between \code{and}/\code{or} modes}

\item{abx_win}{Time-span within which sampling has to occur}

\item{samp_win}{Time-span within which antibiotic administration has to
occur}

\item{si_lwr, si_upr}{Lower/upper extent of SI windows}
}
\description{
The Sepsis-3 label is a combination of a suspected infection and an acute
increase in SOFA score by ≥ 2 points.
}
