% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-si.R
\name{si_data}
\alias{si_data}
\alias{si_windows}
\alias{si}
\title{Suspicion of infection label}
\usage{
si_data(
  source,
  abx_count_win = hours(24L),
  abx_min_count = 1L,
  positive_cultures = FALSE,
  dictionary = read_dictionary("concept-dict"),
  ...
)

si_windows(
  tbl,
  si_mode = c("and", "or"),
  abx_win = hours(24L),
  samp_win = hours(72L),
  si_lwr = hours(48L),
  si_upr = hours(24L)
)

si(source, ...)
}
\arguments{
\item{source}{String valued name of data source}

\item{abx_count_win}{Time span during which to apply the \code{abx_min_count}
criterion}

\item{abx_min_count}{Minimal number of antibiotic administrations}

\item{positive_cultures}{Logical flag indicating whether to require
cultures to be positive}

\item{dictionary}{\code{dictionary} object to use for concept lookup}

\item{...}{Passed to \code{\link[=load_concepts]{load_concepts()}}}

\item{tbl}{\code{ts_tbl} object to use for label computations}

\item{si_mode}{Switch between \code{and}/\code{or} modes}

\item{abx_win}{Time-span within which sampling has to occur}

\item{samp_win}{Time-span within which antibiotic administration has to
occur}

\item{si_lwr, si_upr}{Lower/upper extent of SI windows}
}
\description{
Suspected infection is defined as co-occurrence of of antibiotic treatment and body-fluid sampling.
}
\details{
Suspected infection can occur in one of the two following ways:
\itemize{
\item administration of antibiotics followed by a culture sampling within \code{samp_win} hours
}\preformatted{      samp_win
  |---------------|  
 ABX           sampling (last possible)                
}
\itemize{
\item culture sampling followed by an antibiotic administration within \code{abx_win} hours
}\preformatted{                     abx_win
  |---------------------------------------------|  
sampling                                       ABX (last possible)     
}

The default values of \code{samp_win} and \code{abx_win} are 24 and 72 hours respectively, as per Singer et.al. \url{https://jamanetwork.com/journals/jama/fullarticle/2492881}.

The ealier of the two times (fluid sampling, antibiotic treatment) is taken as the time of suspected infection (SI time). The suspected infection window (SI window) is defined to start \code{si_lwr} hours before the SI time and end \code{si_upr} hours after the SI time. The default values of 48 and 24 hours (respectively) are chosen as used by Seymour et.al. \url{https://jamanetwork.com/journals/jama/fullarticle/2492875} (see Supplemental Material).\preformatted{               48h                       24h
 |------------------------------(|)---------------|   
                               SI time
}

For some datasets, however, information on body fluid sampling is not available for majority of the patients (eICU data). Therefore, an alternative definition of suspected infection is required. For this, we use administration of multiple antibiotics (argument \code{abx_min_count} determines the required number) within \code{abx_count_win} hours. The first time of antibiotic administration is taken as the SI time in this case.
}
\seealso{
\url{http://www.r-project.org}.
}
