% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback-scores.R
\name{sirs_score}
\alias{sirs_score}
\alias{qsofa_score}
\alias{news_score}
\alias{mews_score}
\title{SIRS score label}
\usage{
sirs_score(
  ...,
  win_length = hours(24L),
  keep_components = FALSE,
  interval = NULL
)

qsofa_score(
  ...,
  win_length = hours(24L),
  keep_components = FALSE,
  interval = NULL
)

news_score(
  ...,
  win_length = hours(24L),
  keep_components = FALSE,
  interval = NULL
)

mews_score(
  ...,
  win_length = hours(24L),
  keep_components = FALSE,
  interval = NULL
)
}
\arguments{
\item{...}{Data input used for score
evaluation}

\item{win_length}{Window used for carry forward}

\item{keep_components}{Logical flag indicating whether to return the
individual components alongside the aggregated score}

\item{interval}{Time series interval (only used for checking consistency
of input data)}
}
\description{
The SIRS (Systemic Inflammatory Response Syndrome) score is a commonly used
assessment tool used to track a patient's well-being in an ICU.
}
