% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-sirs.R
\name{sirs_score}
\alias{sirs_score}
\title{SIRS score label}
\usage{
sirs_score(
  temperature,
  heart_rate,
  respiratory_rate,
  pa_co2,
  white_blood_cells,
  bands,
  interval = ricu::interval(temperature)
)
}
\arguments{
\item{temperature, }{heart_rate, respiratory_rate, pa_co2,white_blood_cells, bands Data input used for SIRS score evaluation}

\item{interval}{Time series interval (only used for checking consistency
of input data)}
}
\description{
The SIRS (Systemic Inflammatory Response Syndrome) score is a commonly used
assessment tool used to track a patient's well-being in an ICU.
}
