% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-load.R
\name{load_concepts}
\alias{load_concepts}
\alias{load_concepts.character}
\alias{load_concepts.concept}
\alias{load_concepts.item}
\alias{load_dictionary}
\title{Load concept data}
\usage{
load_concepts(x, ...)

\method{load_concepts}{character}(x, src = NULL, concepts = NULL, ...)

\method{load_concepts}{concept}(
  x,
  aggregate = NA_character_,
  merge_data = TRUE,
  verbose = TRUE,
  ...
)

\method{load_concepts}{item}(
  x,
  patient_ids = NULL,
  id_type = "icustay",
  interval = hours(1L),
  ...
)

load_dictionary(
  source = NULL,
  concepts = NULL,
  dictionary = read_dictionary(source),
  ...
)
}
\arguments{
\item{x}{Object specifying the data to be loaded}

\item{...}{Passed to downstream methods}

\item{src}{A character vector, used to subset the \code{concepts}; \code{NULL}
means no subsetting}

\item{concepts}{Character vector (or NULL, meaning everything) of concept
names}

\item{aggregate}{Controls how data within concepts is aggregated}

\item{merge_data}{Logical flag, specifying whether to merge concepts into
wide format or return a list, each entry corresponding to a concept}

\item{verbose}{Logical flag for muting informational output}

\item{patient_ids}{Optional vector of patient ids to subset the fetched data
with}

\item{id_type}{String specifying the patient id type to return}

\item{interval}{The time interval used to discretize time stamps with,
specified as \code{\link[base:difftime]{base::difftime()}} object}

\item{source}{Corresponds to \code{src}}

\item{dictionary}{Corresponds to \code{concepts}}
}
\description{
Data specified as \code{concept} objects is loaded, aggregated per combination
of id and time-step and merged into wide format.
}
