% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-load.R, R/utils-cli.R
\name{load_concepts}
\alias{load_concepts}
\alias{load_concepts.character}
\alias{load_concepts.concept}
\alias{load_concepts.cncpt}
\alias{load_concepts.num_cncpt}
\alias{load_concepts.fct_cncpt}
\alias{load_concepts.rec_cncpt}
\alias{load_concepts.item}
\alias{load_concepts.col_itm}
\alias{load_concepts.sel_itm}
\alias{load_concepts.rgx_itm}
\alias{load_concepts.hrd_itm}
\alias{load_concepts.fun_itm}
\alias{progress_msg}
\title{Load concept data}
\usage{
load_concepts(x, ...)

\method{load_concepts}{character}(
  x,
  src = NULL,
  concepts = NULL,
  ...,
  dict_name = "concept-dict",
  dict_file = NULL
)

\method{load_concepts}{concept}(x, aggregate = NULL, merge_data = TRUE, verbose = TRUE, ...)

\method{load_concepts}{cncpt}(x, aggregate = NULL, ..., progress = NULL)

\method{load_concepts}{num_cncpt}(x, aggregate = NULL, ..., progress = NULL)

\method{load_concepts}{fct_cncpt}(x, aggregate = NULL, ..., progress = NULL)

\method{load_concepts}{rec_cncpt}(
  x,
  aggregate = NULL,
  patient_ids = NULL,
  id_type = "icustay",
  interval = hours(1L),
  ...,
  progress = NULL
)

\method{load_concepts}{item}(
  x,
  patient_ids = NULL,
  id_type = "icustay",
  interval = hours(1L),
  progress = NULL,
  ...
)

\method{load_concepts}{col_itm}(
  x,
  patient_ids = NULL,
  id_type = "icustay",
  interval = hours(1L),
  ...
)

\method{load_concepts}{sel_itm}(
  x,
  patient_ids = NULL,
  id_type = "icustay",
  interval = hours(1L),
  ...
)

\method{load_concepts}{rgx_itm}(
  x,
  patient_ids = NULL,
  id_type = "icustay",
  interval = hours(1L),
  ...
)

\method{load_concepts}{hrd_itm}(
  x,
  patient_ids = NULL,
  id_type = "icustay",
  interval = hours(1L),
  ...
)

\method{load_concepts}{fun_itm}(
  x,
  patient_ids = NULL,
  id_type = "icustay",
  interval = hours(1L),
  ...
)

progress_msg(...)
}
\arguments{
\item{x}{Object specifying the data to be loaded}

\item{...}{Passed to downstream methods}

\item{src}{A character vector, used to subset the \code{concepts}; \code{NULL}
means no subsetting}

\item{concepts}{The concepts to be used or \code{NULL} in which case
\code{\link[=load_dictionary]{load_dictionary()}} is called}

\item{dict_name, dict_file}{In case not concepts are passed as \code{concepts},
these are forwarded to \code{\link[=load_dictionary]{load_dictionary()}} as \code{name} and \code{file} arguments}

\item{aggregate}{Controls how data within concepts is aggregated}

\item{merge_data}{Logical flag, specifying whether to merge concepts into
wide format or return a list, each entry corresponding to a concept}

\item{verbose}{Logical flag for muting informational output}

\item{progress}{Either \code{NULL}, or a progress bar object as created by
\link[progress:progress_bar]{progress::progress_bar}}

\item{patient_ids}{Optional vector of patient ids to subset the fetched data
with}

\item{id_type}{String specifying the patient id type to return}

\item{interval}{The time interval used to discretize time stamps with,
specified as \code{\link[base:difftime]{base::difftime()}} object}
}
\description{
Data specified as \code{concept} objects is loaded, aggregated per combination
of id and time-step and merged into wide format.
}
