% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{load_src}
\alias{load_src}
\alias{load_src.data_src}
\alias{load_difftime}
\alias{load_difftime.mimic_src}
\alias{load_difftime.mimic_demo_src}
\alias{load_difftime.eicu_src}
\alias{load_difftime.eicu_demo_src}
\alias{load_difftime.hirid_src}
\alias{load_id}
\alias{load_id.data_src}
\alias{load_ts}
\alias{load_ts.data_src}
\alias{data_ts}
\alias{data_id}
\title{Low level functions for loading data}
\usage{
load_src(x, rows, cols = colnames(x))

\method{load_src}{data_src}(x, rows, cols = colnames(x))

load_difftime(x, rows, cols = colnames(x), id_hint = default_col(x, "id"))

\method{load_difftime}{mimic_src}(x, rows, cols = colnames(x), id_hint = default_col(x, "id"))

\method{load_difftime}{mimic_demo_src}(x, rows, cols = colnames(x), id_hint = default_col(x, "id"))

\method{load_difftime}{eicu_src}(x, rows, cols = colnames(x), id_hint = default_col(x, "id"))

\method{load_difftime}{eicu_demo_src}(x, rows, cols = colnames(x), id_hint = default_col(x, "id"))

\method{load_difftime}{hirid_src}(x, rows, cols = colnames(x), id_hint = default_col(x, "id"))

load_id(
  x,
  rows,
  cols = colnames(x),
  id_col = default_col(x, "id"),
  interval = hours(1L)
)

\method{load_id}{data_src}(
  x,
  rows,
  cols = colnames(x),
  id_col = default_col(x, "id"),
  interval = hours(1L)
)

load_ts(
  x,
  rows,
  cols = colnames(x),
  id_col = default_col(x, "id"),
  index_col = default_col(x, "index"),
  interval = hours(1L)
)

\method{load_ts}{data_src}(
  x,
  rows,
  cols = colnames(x),
  id_col = default_col(x, "id"),
  index_col = default_col(x, "index"),
  interval = hours(1L)
)

data_ts(source, table, row_expr, ...)

data_id(source, table, row_expr, ...)
}
\arguments{
\item{x}{Object for which to load data}

\item{rows}{Expression used for row subsetting (NSE)}

\item{cols}{Character vector of column names}

\item{id_hint}{String valued id column selection (not necessarily honored)}

\item{id_col}{The column defining the id of \code{ts_tbl} and \code{id_tbl} objects}

\item{interval}{The time interval used to discretize time stamps with,
specified as \code{\link[base:difftime]{base::difftime()}} object}

\item{index_col}{The column defining the index of \code{ts_tbl} objects}

\item{source}{String specifying the data source}

\item{table}{String specifying the table from which to load data}

\item{row_expr}{Expression to be quoted and used to define a row subsetting;
\code{NULL} corresponds to no row subsetting}

\item{...}{Forwarded to the corresponding function (\code{load_id()} or
\code{load_ts()})}
}
\description{
Depending on the desired return type (\code{ts_tbl}, \code{id_tbl} or plain
\code{data.table}) and on whether the row-subsetting argument is a quoted
expression (suffix \verb{_quo}), this set of functions provides the basic
mechanisms for loading data.
}
