% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{load_src}
\alias{load_src}
\alias{load_src.src_tbl}
\alias{load_src.character}
\alias{load_difftime}
\alias{load_difftime.mimic_tbl}
\alias{load_difftime.eicu_tbl}
\alias{load_difftime.hirid_tbl}
\alias{load_difftime.character}
\alias{load_id}
\alias{load_id.src_tbl}
\alias{load_id.character}
\alias{load_ts}
\alias{load_ts.src_tbl}
\alias{load_ts.character}
\alias{data_id}
\alias{data_ts}
\title{Low level functions for loading data}
\usage{
load_src(x, ...)

\method{load_src}{src_tbl}(x, rows, cols = colnames(x), ...)

\method{load_src}{character}(x, src, ...)

load_difftime(x, ...)

\method{load_difftime}{mimic_tbl}(
  x,
  rows,
  cols = colnames(x),
  id_hint = id_vars(x),
  time_vars = ricu::time_vars(x),
  ...
)

\method{load_difftime}{eicu_tbl}(
  x,
  rows,
  cols = colnames(x),
  id_hint = id_vars(x),
  time_vars = ricu::time_vars(x),
  ...
)

\method{load_difftime}{hirid_tbl}(
  x,
  rows,
  cols = colnames(x),
  id_hint = id_vars(x),
  time_vars = ricu::time_vars(x),
  ...
)

\method{load_difftime}{character}(x, src, ...)

load_id(x, ...)

\method{load_id}{src_tbl}(
  x,
  rows,
  cols = colnames(x),
  id_var = id_vars(x),
  interval = hours(1L),
  time_vars = ricu::time_vars(x),
  ...
)

\method{load_id}{character}(x, src, ...)

load_ts(x, ...)

\method{load_ts}{src_tbl}(
  x,
  rows,
  cols = colnames(x),
  id_var = id_vars(x),
  index_var = ricu::index_var(x),
  interval = hours(1L),
  time_vars = ricu::time_vars(x),
  ...
)

\method{load_ts}{character}(x, src, ...)

data_id(source, table, ...)

data_ts(source, table, ...)
}
\arguments{
\item{x}{Object for which to load data}

\item{...}{Generic consistency}

\item{rows}{Expression used for row subsetting (NSE)}

\item{cols}{Character vector of column names}

\item{src}{Passed to \code{\link[=as_src_tbl]{as_src_tbl()}} in order to determine the data source}

\item{id_hint}{String valued id column selection (not necessarily honored)}

\item{time_vars}{Character vector enumerating the columns to be treated as
timestamps and thus returned as \code{\link[base:difftime]{base::difftime()}} vectors}

\item{id_var}{The column defining the id of \code{ts_tbl} and \code{id_tbl} objects}

\item{interval}{The time interval used to discretize time stamps with,
specified as \code{\link[base:difftime]{base::difftime()}} object}

\item{index_var}{The column defining the index of \code{ts_tbl} objects}

\item{source}{String specifying the data source}

\item{table}{String specifying the table from which to load data}
}
\description{
Depending on the desired return type (\code{ts_tbl}, \code{id_tbl} or plain
\code{data.table}) and on whether the row-subsetting argument is a quoted
expression (suffix \verb{_quo}), this set of functions provides the basic
mechanisms for loading data.
}
