% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricu.R
\docType{package}
\name{ricu-package}
\alias{ricu}
\alias{ricu-package}
\title{ricu: Intensive Care Unit Data with R}
\description{
Focused on (but not exclusive to) data sets hosted on PhysioNet (<https://physionet.org>), 'ricu' provides utilities for download, setup and access of intensive care unit (ICU) data sets. In addition to functions for running arbitrary queries against available data sets, a system for defining clinical concepts and encoding their representations in tabular ICU data is presented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eth-mds/ricu}
  \item \url{https://physionet.org}
  \item Report bugs at \url{https://github.com/eth-mds/ricu/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas.bennett@stat.math.ethz.ch}

Authors:
\itemize{
  \item Drago Plecko \email{drago.plecko@stat.math.ethz.ch}
  \item Ida-Fong Ukor \email{ida-fong.ukor@monashhealth.org}
}

}
\keyword{internal}
