% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{new_src_cfg}
\alias{new_src_cfg}
\alias{is_src_cfg}
\alias{as_src_cfg}
\alias{as_src_cfg.src_cfg}
\alias{as_src_cfg.list}
\alias{src_url}
\title{Data source configuration}
\usage{
new_src_cfg(name, id_cfg, tables, url = NULL)

is_src_cfg(x)

as_src_cfg(x, ...)

\method{as_src_cfg}{src_cfg}(x, ...)

\method{as_src_cfg}{list}(x, ...)

src_url(x)
}
\arguments{
\item{name}{Name of the data source}

\item{id_cfg}{List passed to \code{\link[=new_id_cfg]{new_id_cfg()}}}

\item{tables}{List passed to \code{\link[=as_col_cfg]{as_col_cfg()}} and potentially to
\code{\link[=as_tbl_spec]{as_tbl_spec()}}}

\item{url}{String valued url used for data downloading (see
\code{\link[=download_src]{download_src()}})}

\item{x}{Object to test/coerce}
}
\description{
In order to use a data source with \code{ricu}, a configuration object has to be
created.
}
\keyword{internal}
