% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-class.R, R/config-utils.R
\name{new_src_cfg}
\alias{new_src_cfg}
\alias{is_src_cfg}
\alias{src_url}
\title{Data source configuration}
\usage{
new_src_cfg(
  name,
  id_cfg,
  col_cfg,
  tbl_cfg = NULL,
  url = NULL,
  class_prefix = name
)

is_src_cfg(x)

src_url(x)
}
\arguments{
\item{name}{Name of the data source}

\item{id_cfg}{List passed to \code{\link[=new_id_cfg]{new_id_cfg()}}}

\item{url}{String valued url used for data downloading (see
\code{\link[=download_src]{download_src()}})}

\item{x}{Object to test/coerce}

\item{tables}{List passed to \code{\link[=as_col_cfg]{as_col_cfg()}} and potentially to
\code{\link[=as_tbl_cfg]{as_tbl_cfg()}}}
}
\description{
In order to use a data source with \code{ricu}, a configuration object has to be
created.
}
\keyword{internal}
