% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{stay_windows}
\alias{stay_windows}
\title{Stays}
\usage{
stay_windows(
  x,
  id_type = "icustay",
  win_type = id_type,
  in_time = "start",
  out_time = "end",
  interval = hours(1L)
)
}
\arguments{
\item{x}{Passed to \code{\link[=as_id_cfg]{as_id_cfg()}} and \code{\link[=as_src_env]{as_src_env()}}}

\item{id_type}{Type of ID all returned times are relative to}

\item{win_type}{Type of ID for which the in/out times is returned}

\item{in_time, out_time}{column names of the returned in/out times}

\item{interval}{The time interval used to discretize time stamps with,
specified as \code{\link[base:difftime]{base::difftime()}} object}
}
\value{
An \code{id_tbl} containing the selected IDs and depending on values
passed as \code{in_time} and \code{out_time}, start and end times of the ID passed as
\code{win_var}.
}
\description{
Building on functionality offered by the (internal) function \code{\link[=id_map]{id_map()}},
stay windows as well as (in case of differing values being passed as
\code{id_type} and \code{win_type}) an ID mapping is computed.
}
\seealso{
change_id
}
