% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R, R/meta-id.R, R/meta-tbl.R,
%   R/meta-ts.R, R/meta-utils.R, R/tbl-icu.R
\name{write_psv}
\alias{write_psv}
\alias{read_psv}
\alias{new_id_meta}
\alias{is_id_meta}
\alias{as_id_meta}
\alias{as_id_meta.id_meta}
\alias{as_id_meta.ts_meta}
\alias{meta_cols.id_meta}
\alias{tbl_class.id_meta}
\alias{validate_meta.id_meta}
\alias{new_tbl_index}
\alias{new_tbl_id}
\alias{is_tbl_index}
\alias{is_tbl_id}
\alias{tbl_index.tbl_index}
\alias{tbl_id.tbl_id}
\alias{new_ts_meta}
\alias{is_ts_meta}
\alias{validate_meta.ts_meta}
\alias{is_tbl_meta}
\alias{tbl_meta}
\alias{tbl_meta.tbl_meta}
\alias{tbl_id.tbl_meta}
\alias{validate_meta}
\alias{tbl_meta.icu_tbl}
\alias{tbl_id.icu_tbl}
\alias{tbl_index}
\alias{tbl_id}
\alias{tbl_class}
\alias{tbl_class.icu_tbl}
\title{Read and write utilities}
\usage{
write_psv(x, dir, na_rows = NULL)

read_psv(dir, col_spec = NULL, id_col = "stay_id")

new_id_meta(tbl_id)

is_id_meta(x)

as_id_meta(x)

\method{as_id_meta}{id_meta}(x)

\method{as_id_meta}{ts_meta}(x)

\method{meta_cols}{id_meta}(x)

\method{tbl_class}{id_meta}(x)

\method{validate_meta}{id_meta}(x, meta)

new_tbl_index(index, interval)

new_tbl_id(id, opts = NULL)

is_tbl_index(x)

is_tbl_id(x)

\method{tbl_index}{tbl_index}(x)

\method{tbl_id}{tbl_id}(x)

new_ts_meta(tbl_id, tbl_index)

is_ts_meta(x)

\method{validate_meta}{ts_meta}(x, meta)

is_tbl_meta(x)

tbl_meta(x)

\method{tbl_meta}{tbl_meta}(x)

\method{tbl_id}{tbl_meta}(x)

validate_meta(x, meta)

\method{tbl_meta}{icu_tbl}(x)

\method{tbl_id}{icu_tbl}(x)

tbl_index(x)

tbl_id(x)

tbl_class(x)

\method{tbl_class}{icu_tbl}(x)
}
\arguments{
\item{x}{Object to query/coerce}

\item{dir}{Directory to write the (many) files to or read from}

\item{na_rows}{If \code{TRUE} missing time steps are filled with \code{NaN} values,
if \code{FALSE}, rows where all data columns entries are missing are removed and
if \code{NULL}, data is written as-is}

\item{col_spec}{A column specification as created by \code{\link[readr:cols]{readr::cols()}}}

\item{id_col}{Name of the id column (IDs are generated from file names)}

\item{tbl_id, tbl_index}{Meta data objects to construct \code{tbl_meta} objects
with}

\item{meta}{Meta data object to validate}

\item{index}{Column name of the index column}

\item{interval}{Minimal time interval between rows}

\item{id}{Column name of the ID column}

\item{opts}{Alternative ID systems}
}
\description{
Support for reading from and writing to \code{.psv} files as used for the
Physionet Sepsis Challenge.

Tabular ICU data as used by \code{ricu} frequently contains a column that defines
a grouping structure (e.g. patient IDs, ICU stay IDs, etc.) and a column
holding time information.
}
\references{
Reyna, M., Josef, C., Jeter, R., Shashikumar, S., Moody, B., Westover, M.
B., Sharma, A., Nemati, S., & Clifford, G. (2019). Early Prediction of
Sepsis from Clinical Data -- the PhysioNet Computing in Cardiology
Challenge 2019 (version 1.0.0). PhysioNet.
https://doi.org/10.13026/v64v-d857.
}
\keyword{internal}
