% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-id.R, R/meta-tbl.R, R/meta-ts.R,
%   R/tbl-icu.R, R/tbl-utils.R
\name{rename_cols.id_meta}
\alias{rename_cols.id_meta}
\alias{rename_cols.tbl_index}
\alias{rename_cols.tbl_id}
\alias{rename_cols.ts_meta}
\alias{rm_cols}
\alias{rm_cols.data.table}
\alias{rename_cols}
\alias{rename_cols.data.table}
\alias{rm_cols.icu_tbl}
\alias{rename_cols.icu_tbl}
\alias{is_unique}
\alias{is_unique.default}
\alias{rbind_lst}
\title{Utilities for \code{icu_tbl} objects}
\usage{
\method{rename_cols}{id_meta}(x, new, old, ...)

\method{rename_cols}{tbl_index}(x, new, old, ...)

\method{rename_cols}{tbl_id}(x, new, old, ...)

\method{rename_cols}{ts_meta}(x, new, old, ...)

rm_cols(x, cols, ...)

\method{rm_cols}{data.table}(x, cols, ...)

rename_cols(x, new, old = colnames(x), ...)

\method{rename_cols}{data.table}(x, new, old = colnames(x), ...)

\method{rm_cols}{icu_tbl}(x, cols, ...)

\method{rename_cols}{icu_tbl}(x, new, old = colnames(x), skip_absent = FALSE, ...)

is_unique(x, ...)

\method{is_unique}{default}(x, ...)

rbind_lst(x, ...)
}
\arguments{
\item{x}{Object to operate on}

\item{new}{New column names}

\item{old}{Column names to be replaces by the ones specified as \code{new}}

\item{...}{Generic consistency}

\item{cols}{Column names of columns to remove}

\item{skip_absent}{Silently ignore absent columns in \code{old} (along with
their corresponding replacements in \code{new})}
}
\description{
Utilities for \code{icu_tbl} objects
}
