% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-utils.R
\name{rename_cols}
\alias{rename_cols}
\alias{rm_cols}
\alias{change_interval}
\alias{rm_na}
\alias{dt_gforce}
\alias{is_unique}
\alias{is_unique.default}
\alias{make_unique}
\title{ICU class data utilities}
\usage{
rename_cols(x, new, old = colnames(x), skip_absent = FALSE, by_ref = FALSE)

rm_cols(x, cols, skip_absent = FALSE, by_ref = FALSE)

change_interval(x, new_interval, cols = time_vars(x), by_ref = FALSE)

rm_na(x, cols = data_vars(x), mode = c("all", "any"))

dt_gforce(
  x,
  fun = c("mean", "median", "min", "max", "sum", "prod", "var", "sd", "first", "last"),
  by = meta_vars(x),
  vars = data_vars(x),
  na_rm = !fun \%in\% c("first", "last")
)

is_unique(x, ...)

\method{is_unique}{default}(x, ...)

make_unique(
  x,
  expr = NULL,
  by = meta_vars(x),
  vars = data_vars(x),
  env = NULL,
  ...
)
}
\arguments{
\item{new, old}{Replacement names and existing column names for renaming
columns}

\item{skip_absent}{Logical flag for ignoring non-existent column names}

\item{by_ref}{Logical flag indicating whether to perform the operation by
reference}

\item{cols}{Column names of columns to consider}

\item{new_interval}{Replacement interval length specified as scalar-valued
\code{difftime} object}

\item{mode}{Switch between \code{all} where all entries of a row have to be
missing (for the selected columns) or \code{any}, where a single missing entry
suffices}

\item{fun}{Function name (as string) to apply over groups}

\item{vars}{Column names to apply the function to}

\item{...}{Ignored}

\item{expr}{Expression to apply over groups}

\item{env}{Environment to look up names in \code{expr}}
}
\description{
Several utility functions for working with \code{id_tbl} and \code{ts_tbl} objects
are available, including functions for changing column names, removing
columns, as well as aggregating or removing rows. An important thing to
note is that as \code{id_tbl} (and consequently \code{ts_tbl}) inherits from
\code{data.table}, there are several functions provided by the \code{data.table}
package that are capable of modifying \code{id_tbl} in a way that results in an
object with inconsistent state. An example for this is
\code{\link[data.table:setattr]{data.table::setnames()}}: if an ID column or the index column name is
modified without updating the attribute marking the column as such, this
leads to an invalid object. As \code{\link[data.table:setattr]{data.table::setnames()}} is not an S3
generic function, the only way to control its behavior with respect to
\code{id_tbl} objects is masking the function. As such an approach has its own
down-sides, a separate function, \code{rename_cols()} is provided, which is able
to handle column renaming correctly.
}
\details{
yada yada
}
