% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-class.R, R/tbl-utils.R
\name{unclass_tbl}
\alias{unclass_tbl}
\alias{reclass_tbl}
\alias{as_ptype}
\alias{rename_cols}
\alias{rm_cols}
\alias{change_interval}
\alias{rbind_lst}
\alias{rm_na}
\alias{unmerge}
\alias{dt_gforce}
\alias{is_unique}
\alias{is_unique.default}
\alias{make_unique}
\title{ICU class data utilities}
\usage{
unclass_tbl(x)

reclass_tbl(x, template, stop_on_fail = TRUE)

as_ptype(x)

rename_cols(x, new, old = colnames(x), skip_absent = FALSE, by_ref = FALSE)

rm_cols(x, cols, skip_absent = FALSE, by_ref = FALSE)

change_interval(x, new_interval, cols = time_vars(x), by_ref = FALSE)

rbind_lst(x, ...)

rm_na(x, cols = data_vars(x), mode = c("all", "any"))

unmerge(x, col_groups = as.list(data_vars(x)), by = meta_vars(x), na_rm = TRUE)

dt_gforce(
  x,
  fun = c("mean", "median", "min", "max", "sum", "prod", "var", "sd", "first", "last"),
  by = meta_vars(x),
  vars = data_vars(x),
  na_rm = !fun \%in\% c("first", "last")
)

is_unique(x, ...)

\method{is_unique}{default}(x, ...)

make_unique(
  x,
  expr = NULL,
  by = meta_vars(x),
  vars = data_vars(x),
  env = NULL,
  ...
)
}
\arguments{
\item{x}{Object to modify/query}

\item{template}{Object after which to model the object in question}

\item{stop_on_fail}{Logical flag indicating whether to consider failed
object validation as error}

\item{new, old}{Replacement names and existing column names for renaming
columns}

\item{skip_absent}{Logical flag for ignoring non-existent column names}

\item{by_ref}{Logical flag indicating whether to perform the operation by
reference}

\item{cols}{Column names of columns to consider}

\item{new_interval}{Replacement interval length specified as scalar-valued
\code{difftime} object}

\item{...}{Ignored}

\item{mode}{Switch between \code{all} where all entries of a row have to be
missing (for the selected columns) or \code{any}, where a single missing entry
suffices}

\item{col_groups}{A list of character vectors defining the grouping of
non-by columns}

\item{by}{Columns that will be present in every one of the resulting tables}

\item{na_rm}{Logical flag indicating whether to remove rows that have all
missing entries in the respective \code{col_groups} group}

\item{fun}{Function name (as string) to apply over groups}

\item{vars}{Column names to apply the function to}

\item{expr}{Expression to apply over groups}

\item{env}{Environment to look up names in \code{expr}}
}
\description{
Utilities for working with \code{id_tbl} and \code{ts_tbl} objects.
}
