% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-utils.R
\name{id_vars}
\alias{id_vars}
\alias{id_col}
\alias{index_var}
\alias{index_col}
\alias{meta_vars}
\alias{meta_vars.id_tbl}
\alias{meta_vars.ts_tbl}
\alias{data_vars}
\alias{data_col}
\alias{interval}
\alias{interval.ts_tbl}
\alias{interval.difftime}
\alias{time_unit}
\alias{time_step}
\alias{rename_cols}
\alias{rename_cols.ts_tbl}
\alias{rename_cols.id_tbl}
\alias{rename_cols.data.table}
\alias{rm_cols}
\alias{change_interval}
\alias{change_interval.ts_tbl}
\alias{change_interval.data.table}
\alias{rbind_lst}
\alias{rm_na}
\alias{unmerge}
\alias{dt_gforce}
\alias{is_unique}
\alias{is_unique.default}
\alias{make_unique}
\alias{make_unique_quo}
\title{Utilities for working with id_tbl and ts_tbl objects}
\usage{
id_vars(x)

id_col(x)

index_var(x)

index_col(x)

meta_vars(x)

\method{meta_vars}{id_tbl}(x)

\method{meta_vars}{ts_tbl}(x)

data_vars(x)

data_col(x)

interval(x)

\method{interval}{ts_tbl}(x)

\method{interval}{difftime}(x)

time_unit(x)

time_step(x)

rename_cols(x, new, old = colnames(x), skip_absent = FALSE, by_ref = FALSE)

\method{rename_cols}{ts_tbl}(x, new, old = colnames(x), skip_absent = FALSE, ...)

\method{rename_cols}{id_tbl}(x, new, old = colnames(x), skip_absent = FALSE, ...)

\method{rename_cols}{data.table}(x, new, old = colnames(x), skip_absent = FALSE, by_ref = FALSE)

rm_cols(x, cols, skip_absent = FALSE, by_ref = FALSE)

change_interval(x, new_interval, cols = time_vars(x), by_ref = FALSE)

\method{change_interval}{ts_tbl}(x, new_interval, cols = time_vars(x), ...)

\method{change_interval}{data.table}(x, new_interval, cols = time_vars(x), by_ref = FALSE)

rbind_lst(x, ...)

rm_na(x, cols = data_vars(x), mode = c("all", "any"))

unmerge(x, col_groups = as.list(data_vars(x)), by = meta_vars(x), na_rm = TRUE)

dt_gforce(
  x,
  fun = c("mean", "median", "min", "max", "sum", "prod", "var", "sd", "first", "last"),
  by = meta_vars(x),
  cols = data_vars(x),
  na_rm = !fun \%in\% c("first", "last")
)

is_unique(x, ...)

\method{is_unique}{default}(x, ...)

make_unique(x, expr, fun, ...)

make_unique_quo(x, expr, by = meta_vars(x), cols = data_vars(x), ...)
}
\arguments{
\item{x}{Object to query/operate on}

\item{new, old}{Replacement names and existing column names for renaming
columns}

\item{skip_absent}{Logical flag for ignoring non-existent column names}

\item{by_ref}{Logical flag indicating whether to perform the operation by
reference}

\item{...}{Ignored}

\item{cols}{Column names of columns to consider}

\item{new_interval}{Replacement interval length specified as scalar-valued
\code{difftime} object}

\item{mode}{Switch between \code{all} where all entries of a row have to be
missing (for the selected columns) or \code{any}, where a single missing entry
suffices}

\item{col_groups}{A list of character vectors defining the grouping of
non-by columns}

\item{by}{Columns that will be present in every one of the resulting tables}

\item{na_rm}{Logical flag indicating whether to remove rows that have all
missing entries in the respective \code{col_groups} group}

\item{fun}{Function name (as string) to apply over groups}

\item{expr}{Expression to apply over groups}
}
\description{
Utilities for working with id_tbl and ts_tbl objects
}
