% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tbl.R
\name{as_ts_tbl}
\alias{as_ts_tbl}
\alias{ts_tbl}
\alias{is_ts_tbl}
\alias{[[.ts_tbl}
\alias{[.ts_tbl}
\alias{$.ts_tbl}
\alias{dim.ts_tbl}
\alias{length.ts_tbl}
\alias{dimnames.ts_tbl}
\alias{names.ts_tbl}
\alias{time_col}
\alias{id_cols}
\alias{val_cols}
\alias{step_size}
\alias{step_time}
\alias{time_unit}
\alias{time_unit<-}
\alias{as.data.table.ts_tbl}
\title{Methods for creating and inspecting ts_tbl objects}
\usage{
as_ts_tbl(tbl, id_cols, time_col = NULL, step_size = 1L)

ts_tbl(..., id_cols, time_col = NULL, step_size = 1L)

is_ts_tbl(x)

\method{[[}{ts_tbl}(x, ...)

\method{[}{ts_tbl}(x, ...)

\method{$}{ts_tbl}(x, ...)

\method{dim}{ts_tbl}(x)

\method{length}{ts_tbl}(x)

\method{dimnames}{ts_tbl}(x)

\method{names}{ts_tbl}(x)

time_col(x)

id_cols(x)

val_cols(x)

step_size(x)

step_time(x)

time_unit(x)

time_unit(x) <- value

\method{as.data.table}{ts_tbl}(x, ...)
}
\arguments{
\item{tbl}{An object inheriting from \code{data.frame}.}

\item{id_cols}{Character or numeric vector of at least length 1, identifying
columns that combined with time stamps, uniquely identify rows.}

\item{time_col}{Character or numeric vector of length 1, identifying the
column that defines temporal ordering.}

\item{step_size}{Scalar value, defining time steps between rows (assuming
complete time series data).}

\item{...}{Passed to \code{\link[data.table:data.table]{data.table::data.table()}}/generic compatibility.}

\item{x}{A \code{ts_tbl} object.}

\item{value}{New time unit.}
}
\description{
Methods for creating and inspecting ts_tbl objects
}
