% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tbl.R
\name{as_ts_tbl}
\alias{as_ts_tbl}
\alias{ts_tbl}
\alias{is_ts_tbl}
\alias{[.ts_tbl}
\alias{dimnames.ts_tbl}
\alias{print.ts_tbl}
\alias{format.ts_tbl}
\alias{tbl_sum.ts_tbl}
\alias{str.ts_tbl}
\alias{ts_index}
\alias{ts_key}
\alias{by_cols}
\alias{val_cols}
\alias{ts_step}
\alias{step_time}
\alias{time_unit}
\alias{time_unit<-}
\alias{as.data.table.ts_tbl}
\alias{as.data.frame.ts_tbl}
\title{Methods for creating and inspecting ts_tbl objects}
\usage{
as_ts_tbl(tbl, ts_key, ts_index = NULL, ts_step = 1L)

ts_tbl(..., ts_key, ts_index = NULL, ts_step = 1L)

is_ts_tbl(x)

\method{[}{ts_tbl}(x, ...)

\method{dimnames}{ts_tbl}(x)

\method{print}{ts_tbl}(x, ..., n = NULL, width = NULL,
  n_extra = NULL)

\method{format}{ts_tbl}(x, ..., n = NULL, width = NULL,
  n_extra = NULL)

\method{tbl_sum}{ts_tbl}(x)

\method{str}{ts_tbl}(object, ...)

ts_index(x)

ts_key(x)

by_cols(x)

val_cols(x)

ts_step(x)

step_time(x)

time_unit(x)

time_unit(x) <- value

\method{as.data.table}{ts_tbl}(x, ...)

\method{as.data.frame}{ts_tbl}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{tbl}{An object inheriting from \code{data.frame}.}

\item{ts_key}{Character or numeric vector of at least length 1, identifying
columns that combined with time stamps, uniquely identify rows.}

\item{ts_index}{Character or numeric vector of length 1, identifying the
column that defines temporal ordering.}

\item{ts_step}{Scalar value, defining time steps between rows (assuming
complete time series data).}

\item{...}{Passed to \code{\link[data.table:data.table]{data.table::data.table()}}/generic compatibility.}

\item{x}{A \code{ts_tbl} object.}

\item{object}{A \code{ts_tbl} object.}

\item{value}{New time unit.}

\item{row.names, optional}{Generic consistency: passing anything other than
the default value issues a warning.}
}
\description{
Methods for creating and inspecting ts_tbl objects
}
