% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tbl.R
\name{as_ts_tbl}
\alias{as_ts_tbl}
\alias{ts_tbl}
\alias{is_ts_tbl}
\alias{[.ts_tbl}
\alias{dimnames.ts_tbl}
\alias{print.ts_tbl}
\alias{format.ts_tbl}
\alias{tbl_sum.ts_tbl}
\alias{str.ts_tbl}
\alias{ts_index}
\alias{ts_key}
\alias{by_cols}
\alias{val_cols}
\alias{ts_step}
\alias{step_time}
\alias{time_unit}
\alias{time_unit<-}
\alias{as.data.table.ts_tbl}
\alias{as.data.frame.ts_tbl}
\alias{merge.ts_tbl}
\alias{rbind.ts_tbl}
\alias{unique.ts_tbl}
\alias{is_unique}
\title{Methods for creating and inspecting ts_tbl objects}
\usage{
as_ts_tbl(tbl, ts_key, ts_index = NULL, ts_step = 1L)

ts_tbl(..., ts_key, ts_index = NULL, ts_step = 1L)

is_ts_tbl(x)

\method{[}{ts_tbl}(x, ...)

\method{dimnames}{ts_tbl}(x)

\method{print}{ts_tbl}(x, ..., n = NULL, width = NULL,
  n_extra = NULL)

\method{format}{ts_tbl}(x, ..., n = NULL, width = NULL,
  n_extra = NULL)

\method{tbl_sum}{ts_tbl}(x)

\method{str}{ts_tbl}(object, ...)

ts_index(x)

ts_key(x)

by_cols(x)

val_cols(x)

ts_step(x)

step_time(x)

time_unit(x)

time_unit(x) <- value

\method{as.data.table}{ts_tbl}(x, ...)

\method{as.data.frame}{ts_tbl}(x, row.names = NULL, optional = FALSE,
  ...)

\method{merge}{ts_tbl}(x, y, by = by_cols(x), ...)

\method{rbind}{ts_tbl}(x, ..., use.names = "check", fill = FALSE,
  idcol = NULL)

\method{unique}{ts_tbl}(x, ...)

is_unique(x, by = by_cols(x))
}
\arguments{
\item{tbl}{An object inheriting from \code{data.frame}.}

\item{ts_key}{Character or numeric vector of at least length 1, identifying
columns that combined with time stamps, uniquely identify rows.}

\item{ts_index}{Character or numeric vector of length 1, identifying the
column that defines temporal ordering.}

\item{ts_step}{Scalar value, defining time steps between rows (assuming
complete time series data).}

\item{...}{Passed to \code{\link[data.table:data.table]{data.table::data.table()}}/generic compatibility.}

\item{x}{A \code{ts_tbl} object.}

\item{object}{A \code{ts_tbl} object.}

\item{value}{New time unit.}

\item{row.names, optional}{Generic consistency: passing anything other than
the default value issues a warning.}

\item{y}{A \code{ts_tbl} object.}

\item{use.names}{\code{TRUE} binds by matching column name, \code{FALSE} by position. `check` (default) warns if all items don't have the same names in the same order and then currently proceeds as if `use.names=FALSE` for backwards compatibility (\code{TRUE} in future); see news for v1.12.2.}

\item{fill}{\code{TRUE} fills missing columns with NAs. By default \code{FALSE}. When \code{TRUE}, \code{use.names} is set to \code{TRUE}.}

\item{idcol}{Creates a column in the result showing which list item those rows came from. \code{TRUE} names this column \code{".id"}. \code{idcol="file"} names this column \code{"file"}. If the input list has names, those names are the values placed in this id column, otherwise the values are an integer vector \code{1:length(l)}. See \code{examples}.}
}
\description{
Methods for creating and inspecting ts_tbl objects
}
