% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-base.R, R/ts-tbl.R
\name{[.ts_tbl}
\alias{[.ts_tbl}
\alias{dimnames.ts_tbl}
\alias{print.ts_tbl}
\alias{format.ts_tbl}
\alias{tbl_sum.ts_tbl}
\alias{str.ts_tbl}
\alias{as.data.table.ts_tbl}
\alias{as.data.frame.ts_tbl}
\alias{ts_tbl}
\alias{as_ts_tbl}
\alias{is_ts_tbl}
\title{Methods for creating and inspecting ts_tbl objects}
\usage{
\method{[}{ts_tbl}(x, ...)

\method{dimnames}{ts_tbl}(x)

\method{print}{ts_tbl}(x, ..., n = NULL, width = NULL, n_extra = NULL)

\method{format}{ts_tbl}(x, ..., n = NULL, width = NULL, n_extra = NULL)

\method{tbl_sum}{ts_tbl}(x)

\method{str}{ts_tbl}(object, ...)

\method{as.data.table}{ts_tbl}(x, ...)

\method{as.data.frame}{ts_tbl}(x, row.names = NULL, optional = FALSE, ...)

ts_tbl(..., key, index = NULL, interval = hours(1L))

as_ts_tbl(tbl, key, index = NULL, interval = hours(1L))

is_ts_tbl(x)
}
\arguments{
\item{x}{A \code{ts_tbl} object.}

\item{...}{Passed to \code{\link[data.table:data.table]{data.table::data.table()}}/generic compatibility.}

\item{object}{A \code{ts_tbl} object.}

\item{row.names, optional}{Generic consistency: passing anything other than
the default value issues a warning.}

\item{key}{Character or numeric vector of at least length 1, identifying
columns that combined with time stamps, uniquely identify rows.}

\item{index}{Character or numeric vector of length 1, identifying the
column that defines temporal ordering.}

\item{interval}{Scalar value, defining time steps between rows (assuming
complete time series data).}

\item{tbl}{An object inheriting from \code{data.frame}.}
}
\description{
Methods for creating and inspecting ts_tbl objects
}
