% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-utils.R
\name{expand_limits}
\alias{expand_limits}
\alias{has_no_gaps}
\alias{has_gaps}
\alias{fill_gaps}
\alias{slide}
\alias{slide_quo}
\alias{slide_index}
\alias{slide_index_quo}
\alias{hop}
\alias{hop_quo}
\title{Time series utility functions}
\usage{
expand_limits(
  x,
  min_col = "min",
  max_col = "max",
  step_size = 1L,
  id_cols = NULL,
  new_col = "hadm_time"
)

has_no_gaps(x)

has_gaps(...)

fill_gaps(x, limits = NULL, ...)

slide(x, expr, ...)

slide_quo(x, expr, before, after = hours(0L), ...)

slide_index(x, expr, ...)

slide_index_quo(x, expr, index, before, after = hours(0L), ...)

hop(x, expr, ...)

hop_quo(
  x,
  expr,
  windows,
  full_window = FALSE,
  lwr_col = "min_time",
  upr_col = "max_time"
)
}
\arguments{
\item{x}{\code{ts_tbl} object to use}

\item{min_col, max_col}{Name of the columns that represent lower and upper
windows bounds}

\item{step_size}{Controls the step size used to interpolate between
\code{min_col} and \code{max_col}}

\item{id_cols}{Names of the id columns}

\item{new_col}{Name of the new index column}

\item{...}{Passed to the \verb{*_quo} version and from there to \code{hop_quo()}}

\item{limits}{A table with columns for lower and upper window bounds}

\item{expr}{Expression (quoted for \verb{*_quo} and unquoted otherwise) to be
evaluated over each window}

\item{before, after}{Time span to look back/forward}

\item{index}{A vector of times around which windows are spanned (relative
to the index)}

\item{windows}{An \code{icu_tbl} defining the windows to span}

\item{full_window}{Logical flag controlling how the situation is handled
where the sliding window extends beyond available data}

\item{lwr_col, upr_col}{Names of columns (in \code{windows}) of lower/upper
window bounds}
}
\description{
Time series utility functions
}
