% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_or_na}
\alias{min_or_na}
\alias{max_or_na}
\alias{sum_or_na}
\alias{is_val}
\alias{is_true}
\alias{last_elem}
\alias{first_elem}
\alias{replace_na}
\title{Utility functions}
\usage{
min_or_na(x)

max_or_na(x)

sum_or_na(x)

is_val(x, val)

is_true(x)

last_elem(x)

first_elem(x)

replace_na(x, val)
}
\arguments{
\item{x}{Object to test}
}
\description{
Various utility functions
}
